% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{check_config}
\alias{check_config}
\title{Check config}
\usage{
check_config(config)
}
\description{
Checks that config is valid.
}
\keyword{#}
\keyword{>>}
\keyword{and}
\keyword{database$apiKey,}
\keyword{database$databaseURL}
\keyword{database$projectId}
\keyword{fill}
\keyword{firebase.yml}
\keyword{in}
\keyword{internal}
