% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_crew.R
\name{class_crew}
\alias{class_crew}
\alias{crew}
\title{Crew class.}
\description{
\code{R6} class for a crew.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Character of length 1, crew name.}

\item{\code{store}}{\code{R6} store object.}

\item{\code{worker_definition}}{\code{R6ClassGenerator}
object with the worker definition.}

\item{\code{worker_list}}{Named list of worker objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{class_crew$new()}}
\item \href{#method-validate}{\code{class_crew$validate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Crew constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_crew$new(
  name = basename(tempfile(pattern = "crew_")),
  store = class_store_local$new(),
  worker_definition = class_worker,
  worker_list = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character of length 1, crew name.}

\item{\code{store}}{\code{R6} store object.}

\item{\code{worker_definition}}{\code{R6ClassGenerator}
object with the worker definition.}

\item{\code{worker_list}}{Named list of \code{R6} worker objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Crew validator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_crew$validate()}\if{html}{\out{</div>}}
}

}
}
