% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_queue.R
\name{class_queue}
\alias{class_queue}
\title{Task queue class}
\description{
A task queue for short-lived worker management tasks
such as launching and polling.

Adapted from
\url{https://github.com/r-lib/callr/blob/811a02f604de2cf03264f6b35ce9ec8a412f2581/vignettes/taskq.R} # nolint
under the MIT license. See also the \code{crew} package \code{NOTICE} file.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tasks}}{Data frame of tasks.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{class_queue$new()}}
\item \href{#method-push}{\code{class_queue$push()}}
\item \href{#method-pop}{\code{class_queue$pop()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Task queue constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_queue$new(workers = 4L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workers}}{Number of local processes that
iterate through tasks.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{new()} method calls the constructor
and returns a task queue object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-push"></a>}}
\if{latex}{\out{\hypertarget{method-push}{}}}
\subsection{Method \code{push()}}{
Push a task to the queue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_queue$push(fun, args = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{Function to run for the task.}

\item{\code{args}}{Arguments to \code{fun}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invsibly)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pop"></a>}}
\if{latex}{\out{\hypertarget{method-pop}{}}}
\subsection{Method \code{pop()}}{
Poll the workers and pop a result off the queue
if available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_queue$pop(timeout = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeout}}{Number of seconds of timeout for polling.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The result of a done task if available, \code{NULL} if there
are no done tasks. The return value is a list with the return
value of the function and the task ID.
}
}
}
