% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_store.R
\name{class_store}
\alias{class_store}
\alias{store}
\title{Store class.}
\description{
\code{R6} class for a data store.
}
\details{
A data store is how a worker object sends jobs to its
underlying R process and receives output back. Subclasses may
use the local file system or cloud storage (e.g. Amazon S3 buckets).
A store object is part of a crew.
}
\examples{
store <- class_store_local$new()
store$write_output("worker_name", list(value = "job_output"))
store$read_output("worker_name")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dir_root}}{Character of length 1, root file path or prefix
where the files are located.}

\item{\code{dir_input}}{Character of length 1, worker input directory.}

\item{\code{dir_output}}{Character of length 1, worker output directory.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{class_store$new()}}
\item \href{#method-path_input}{\code{class_store$path_input()}}
\item \href{#method-path_output}{\code{class_store$path_output()}}
\item \href{#method-validate}{\code{class_store$validate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Store constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_store$new(dir_root = tempfile())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir_root}}{Character of length 1, file path or prefix
where all the files are located.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{new()} method calls the constructor and returns
a new store object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-path_input"></a>}}
\if{latex}{\out{\hypertarget{method-path_input}{}}}
\subsection{Method \code{path_input()}}{
Path to job input.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_store$path_input(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Worker name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character of length 1, path to job input for a given worker.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-path_output"></a>}}
\if{latex}{\out{\hypertarget{method-path_output}{}}}
\subsection{Method \code{path_output()}}{
Path to job output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_store$path_output(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Worker name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character of length 1, path to job output for a given worker.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Store validator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_store$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
}
