% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_store_local.R
\name{class_store_local}
\alias{class_store_local}
\title{Local store class.}
\description{
\code{R6} class for a local store.
}
\section{Super class}{
\code{\link[crew:store]{crew::store}} -> \code{store_local}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-write}{\code{class_store_local$write()}}
\item \href{#method-read}{\code{class_store_local$read()}}
\item \href{#method-exists}{\code{class_store_local$exists()}}
\item \href{#method-delete}{\code{class_store_local$delete()}}
\item \href{#method-destroy}{\code{class_store_local$destroy()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="store" data-id="initialize">}\href{../../crew/html/store.html#method-initialize}{\code{crew::store$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="store" data-id="path">}\href{../../crew/html/store.html#method-path}{\code{crew::store$path()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="store" data-id="path_temp">}\href{../../crew/html/store.html#method-path_temp}{\code{crew::store$path_temp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="store" data-id="validate">}\href{../../crew/html/store.html#method-validate}{\code{crew::store$validate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write"></a>}}
\if{latex}{\out{\hypertarget{method-write}{}}}
\subsection{Method \code{write()}}{
Write worker data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_store_local$write(name, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character of length 1, worker name.}

\item{\code{data}}{Data to write.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
Read worker output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_store_local$read(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character of length 1, worker name.}

\item{\code{output}}{Worker output to write.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exists"></a>}}
\if{latex}{\out{\hypertarget{method-exists}{}}}
\subsection{Method \code{exists()}}{
Exists worker data?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_store_local$exists(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character of length 1, worker name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Delete worker data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_store_local$delete(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character of length 1, worker name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-destroy"></a>}}
\if{latex}{\out{\hypertarget{method-destroy}{}}}
\subsection{Method \code{destroy()}}{
Delete all the files in the data store.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_store_local$destroy()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character of length 1, worker name}
}
\if{html}{\out{</div>}}
}
}
}
