% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_worker_callr.R
\name{class_worker_callr}
\alias{class_worker_callr}
\alias{worker_callr}
\title{\code{callr} worker class.}
\description{
\code{R6} class for a \code{callr} worker.
}
\section{Super class}{
\code{\link[crew:worker]{crew::worker}} -> \code{worker_callr}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{process}}{A \code{callr::r_bg()} process handler.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-launch}{\code{class_worker_callr$launch()}}
\item \href{#method-alive}{\code{class_worker_callr$alive()}}
\item \href{#method-shutdown}{\code{class_worker_callr$shutdown()}}
\item \href{#method-validate}{\code{class_worker_callr$validate()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="done">}\href{../../crew/html/worker.html#method-done}{\code{crew::worker$done()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="initialize">}\href{../../crew/html/worker.html#method-initialize}{\code{crew::worker$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="receive">}\href{../../crew/html/worker.html#method-receive}{\code{crew::worker$receive()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="send">}\href{../../crew/html/worker.html#method-send}{\code{crew::worker$send()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-launch"></a>}}
\if{latex}{\out{\hypertarget{method-launch}{}}}
\subsection{Method \code{launch()}}{
Launch the worker.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker_callr$launch()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-alive"></a>}}
\if{latex}{\out{\hypertarget{method-alive}{}}}
\subsection{Method \code{alive()}}{
\code{TRUE} if the worker is alive and \code{FALSE} otherwise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker_callr$alive()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shutdown"></a>}}
\if{latex}{\out{\hypertarget{method-shutdown}{}}}
\subsection{Method \code{shutdown()}}{
Gracefully shut down the worker.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker_callr$shutdown()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Worker validator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker_callr$validate()}\if{html}{\out{</div>}}
}

}
}
