% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_package.R
\docType{package}
\name{crew-package}
\alias{crew-package}
\title{crew: a distributed worker launcher}
\description{
In computationally demanding analysis workflows,
statisticians and data scientists asynchronously deploy
long-running tasks to distributed systems, ranging from
traditional clusters to cloud services. The \code{crew} package
extends existing task schedulers to run workers on these systems.
With its unifying interface to multiple backends,
\code{crew} resembles packages \code{clustermq},
\code{future}, and \code{batchtools}.
}
\concept{help}
