% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_launcher_callr.R
\name{crew_class_launcher_callr}
\alias{crew_class_launcher_callr}
\title{\code{mirai} launcher class}
\description{
\code{R6} class to launch and manage \code{mirai} workers.
}
\details{
See \code{\link[=crew_launcher_callr]{crew_launcher_callr()}}.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_router()
router$connect()
launcher <- crew_launcher_callr()
launcher$populate(sockets = router$sockets_listening())
launcher$running() # 0
launcher$launch()
launcher$running() # 1
launcher$terminate()
launcher$running() # 0
router$disconnect()
}
}
\seealso{
Other launchers: 
\code{\link{crew_class_launcher}},
\code{\link{crew_launcher_callr}()}
}
\concept{launchers}
\section{Super class}{
\code{\link[crew:crew_class_launcher]{crew::crew_class_launcher}} -> \code{crew_class_launcher_callr}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{processes}}{List of \code{callr} process handles.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_launcher_callr-validate}{\code{crew_class_launcher_callr$validate()}}
\item \href{#method-crew_class_launcher_callr-launch}{\code{crew_class_launcher_callr$launch()}}
\item \href{#method-crew_class_launcher_callr-terminate}{\code{crew_class_launcher_callr$terminate()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_class_launcher" data-id="initialize"><a href='../../crew/html/crew_class_launcher.html#method-crew_class_launcher-initialize'><code>crew::crew_class_launcher$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher_callr-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher_callr-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the launcher.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher_callr$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher_callr-launch"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher_callr-launch}{}}}
\subsection{Method \code{launch()}}{
Launch one or more workers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher_callr$launch(sockets = character(0))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sockets}}{Sockets where the workers will dial in.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The actual number of newly launched workers
is less than or equal to the number of sockets
supplied to \verb{$populate()} minus the number of workers
already running.
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher_callr-terminate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher_callr-terminate}{}}}
\subsection{Method \code{terminate()}}{
Terminate all workers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher_callr$terminate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
}
