% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_mirai_controller.R
\name{crew_class_mirai_controller}
\alias{crew_class_mirai_controller}
\title{\code{mirai} controller class}
\description{
\code{R6} class for \code{mirai} controllers.
}
\details{
See \code{\link[=crew_mirai_controller]{crew_mirai_controller()}}.
}
\seealso{
Other mirai: 
\code{\link{crew_class_mirai_launcher_callr}},
\code{\link{crew_class_mirai_router}},
\code{\link{crew_mirai_controller}()},
\code{\link{crew_mirai_launcher_callr}()},
\code{\link{crew_mirai_router}()}
}
\concept{mirai}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{router}}{Router object.}

\item{\code{launcher}}{Launcher object.}

\item{\code{scale_method}}{Scaling method. See \code{\link[=crew_mirai_controller]{crew_mirai_controller()}}.}

\item{\code{queue}}{List of tasks in the queue.}

\item{\code{results}}{List of finished tasks}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_mirai_controller-new}{\code{crew_class_mirai_controller$new()}}
\item \href{#method-crew_class_mirai_controller-validate}{\code{crew_class_mirai_controller$validate()}}
\item \href{#method-crew_class_mirai_controller-connect}{\code{crew_class_mirai_controller$connect()}}
\item \href{#method-crew_class_mirai_controller-launch}{\code{crew_class_mirai_controller$launch()}}
\item \href{#method-crew_class_mirai_controller-collect}{\code{crew_class_mirai_controller$collect()}}
\item \href{#method-crew_class_mirai_controller-scale}{\code{crew_class_mirai_controller$scale()}}
\item \href{#method-crew_class_mirai_controller-push}{\code{crew_class_mirai_controller$push()}}
\item \href{#method-crew_class_mirai_controller-pop}{\code{crew_class_mirai_controller$pop()}}
\item \href{#method-crew_class_mirai_controller-wait}{\code{crew_class_mirai_controller$wait()}}
\item \href{#method-crew_class_mirai_controller-terminate}{\code{crew_class_mirai_controller$terminate()}}
\item \href{#method-crew_class_mirai_controller-clone}{\code{crew_class_mirai_controller$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-new}{}}}
\subsection{Method \code{new()}}{
\code{mirai} controller constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$new(
  router = NULL,
  launcher = NULL,
  scale_method = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{router}}{Router object. See \code{\link[=crew_mirai_controller]{crew_mirai_controller()}}.}

\item{\code{launcher}}{Launcher object. See \code{\link[=crew_mirai_controller]{crew_mirai_controller()}}.}

\item{\code{scale_method}}{Scaling method. See \code{\link[=crew_mirai_controller]{crew_mirai_controller()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{R6} object with the controller object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the router.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-connect"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-connect}{}}}
\subsection{Method \code{connect()}}{
Connect the router and register the TCP sockets
with the launcher.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$connect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-launch"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-launch}{}}}
\subsection{Method \code{launch()}}{
Launch one or more workers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$launch(n = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Maximum number of workers to launch.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The actual number of newly launched workers
is capped at the max set when the controller
was created. Unless you are launching an array of
persistent workers or have \code{scale_method = "none"},
this method does not usually need to be called
directly because \code{push(scale = TRUE)} already launches
workers depending on \code{scale_method}.
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-collect"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-collect}{}}}
\subsection{Method \code{collect()}}{
Check for done tasks and move the results to
the results list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$collect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly). Removes elements from the \code{queue}
list as applicable and moves them to the \code{results} list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-scale"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-scale}{}}}
\subsection{Method \code{scale()}}{
Automatically scale up the number of workers if needed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$scale()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method is called during \code{push()}, and the method for
scaling up workers is governed by the \code{scale_method}
argument of \code{\link[=crew_mirai_controller]{crew_mirai_controller()}}. It is not meant to be called
manually. If called manually, it is recommended to call \code{collect()}
first so \code{scale()} can accurately assess the demand.
For finer control of the number of workers launched,
call \code{launch()} on the controller with the exact desired
number of workers.
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-push"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-push}{}}}
\subsection{Method \code{push()}}{
Push a task to the head of the task list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$push(
  command,
  args = list(),
  timeout = NULL,
  name = NULL,
  scale = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{command}}{R code with the task to run.}

\item{\code{args}}{A list of objects referenced in \code{expr}.}

\item{\code{timeout}}{Time in milliseconds for the task.}

\item{\code{name}}{Optional name of the task. Replaced with a random name
if \code{NULL} or in conflict with an existing name in the task list.}

\item{\code{scale}}{Logical, whether to automatically scale workers to meet
demand. If \code{TRUE}, then \code{collect()} runs first so demand can be
properly assessed before scaling.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-pop"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-pop}{}}}
\subsection{Method \code{pop()}}{
Pop a completed task from the results data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$pop(collect = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collect}}{Whether to run the \code{collect()} method to collect all
available results before calling \code{pop()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If there is a completed task available to collect, the return
value is a one-row data frame with the results, warnings, and errors.
Otherwise, if there are no results available to collect,
the return value is \code{NULL}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-wait"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-wait}{}}}
\subsection{Method \code{wait()}}{
Wait for a result to be ready.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$wait(timeout = Inf, wait = 0.1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeout}}{Timeout length in seconds waiting for
results to become available.}

\item{\code{wait}}{Number of seconds to wait between polling intervals
while checking for results.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invisibly). Call \code{pop()} to get the result.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-terminate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-terminate}{}}}
\subsection{Method \code{terminate()}}{
Terminate the workers and disconnect the router.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$terminate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-clone"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
