% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_mirai_launcher_callr.R
\name{crew_class_mirai_launcher_callr}
\alias{crew_class_mirai_launcher_callr}
\title{\code{mirai} launcher class}
\description{
\code{R6} class to launch and manage \code{mirai} workers.
}
\details{
See \code{\link[=crew_mirai_launcher_callr]{crew_mirai_launcher_callr()}}.
}
\seealso{
Other launchers: 
\code{\link{crew_mirai_launcher_callr}()}
}
\concept{launchers}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sockets}}{TCP sockets for listening to the workers.}

\item{\code{workers}}{List of \code{callr::r_session()} handles for workers.
The handle is \code{NA} if it is not yet called.}

\item{\code{seconds_idle}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{seconds_wall_time}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{seconds_poll_high}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{seconds_poll_low}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{max_tasks}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{async_dial}}{Argument to \code{crew_mirai_launcher_callr()}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_mirai_launcher_callr-new}{\code{crew_class_mirai_launcher_callr$new()}}
\item \href{#method-crew_class_mirai_launcher_callr-finalize}{\code{crew_class_mirai_launcher_callr$finalize()}}
\item \href{#method-crew_class_mirai_launcher_callr-validate}{\code{crew_class_mirai_launcher_callr$validate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_launcher_callr-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_launcher_callr-new}{}}}
\subsection{Method \code{new()}}{
\code{mirai} launcher constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_launcher_callr$new(
  sockets = NULL,
  workers = NULL,
  seconds_idle = NULL,
  seconds_wall_time = NULL,
  seconds_poll_high = NULL,
  seconds_poll_low = NULL,
  max_tasks = NULL,
  async_dial = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sockets}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{workers}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{seconds_idle}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{seconds_wall_time}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{seconds_poll_high}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{seconds_poll_low}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{max_tasks}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{async_dial}}{Argument to \code{crew_mirai_launcher_callr()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{R6} object with the launcher.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_launcher_callr-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_launcher_callr-finalize}{}}}
\subsection{Method \code{finalize()}}{
Disconnect at garbage collection time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_launcher_callr$finalize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_launcher_callr-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_launcher_callr-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the launcher.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_launcher_callr$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
}
