% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_stage.R
\name{crew_class_stage}
\alias{crew_class_stage}
\title{\code{R6} stage class.}
\description{
\code{R6} class for stage configuration.
}
\details{
See \code{\link[=crew_stage]{crew_stage()}}.
}
\examples{
crew_stage()
}
\seealso{
Other class: 
\code{\link{crew_class_async}},
\code{\link{crew_class_client}},
\code{\link{crew_class_controller_group}},
\code{\link{crew_class_controller}},
\code{\link{crew_class_launcher}},
\code{\link{crew_class_tls}}
}
\concept{class}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{condition}}{\code{nanonext} condition variable that inherits
signals from \code{mirai::nextget("cv")}.}

\item{\code{unpopped}}{Number of observed unpopped tasks.}

\item{\code{popped}}{Number of observed popped tasks.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_stage-validate}{\code{crew_class_stage$validate()}}
\item \href{#method-crew_class_stage-start}{\code{crew_class_stage$start()}}
\item \href{#method-crew_class_stage-inherit}{\code{crew_class_stage$inherit()}}
\item \href{#method-crew_class_stage-resolved}{\code{crew_class_stage$resolved()}}
\item \href{#method-crew_class_stage-pop}{\code{crew_class_stage$pop()}}
\item \href{#method-crew_class_stage-wait_unobserved}{\code{crew_class_stage$wait_unobserved()}}
\item \href{#method-crew_class_stage-wait_unpopped}{\code{crew_class_stage$wait_unpopped()}}
\item \href{#method-crew_class_stage-wait_resolved}{\code{crew_class_stage$wait_resolved()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_stage-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_stage-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_stage$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_stage-start"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_stage-start}{}}}
\subsection{Method \code{start()}}{
Initialize the stage object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_stage$start()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_stage-inherit"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_stage-inherit}{}}}
\subsection{Method \code{inherit()}}{
Inherit signals from a different condition variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_stage$inherit(condition)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{condition}}{A \code{nanonext} condition variable which will forward
signals to \code{self$condition}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_stage-resolved"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_stage-resolved}{}}}
\subsection{Method \code{resolved()}}{
Count the total count of resolved tasks.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_stage$resolved()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Positive integer of length 1, total number of resolved tasks.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_stage-pop"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_stage-pop}{}}}
\subsection{Method \code{pop()}}{
Register a popped task.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_stage$pop()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_stage-wait_unobserved"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_stage-wait_unobserved}{}}}
\subsection{Method \code{wait_unobserved()}}{
Wait until an unobserved task resolves or the timeout
is reached.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_stage$wait_unobserved(seconds_timeout = Inf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seconds_timeout}}{Positive numeric of length 1,
Number of seconds to wait before timing out.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_stage-wait_unpopped"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_stage-wait_unpopped}{}}}
\subsection{Method \code{wait_unpopped()}}{
Wait until at least one task is available to be popped.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_stage$wait_unpopped(seconds_timeout = Inf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seconds_timeout}}{Positive numeric of length 1,
Number of seconds to wait before timing out.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_stage-wait_resolved"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_stage-wait_resolved}{}}}
\subsection{Method \code{wait_resolved()}}{
Wait for until the number of
resolved tasks reaches a specified number.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_stage$wait_resolved(seconds_timeout = 0.25, resolved = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seconds_timeout}}{Positive numeric of length 1,
Number of seconds to wait before timing out.}

\item{\code{resolved}}{Positive integer of length 1. This method waits
until the number of resolved tasks reaches this value or above.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invisibly).
}
}
}
