% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_throttle.R
\name{crew_class_throttle}
\alias{crew_class_throttle}
\title{\code{R6} throttle class.}
\description{
\code{R6} class for throttle configuration.
}
\details{
See \code{\link[=crew_throttle]{crew_throttle()}}.
}
\examples{
throttle <- crew_throttle(seconds_interval = 0.5)
throttle$poll()
throttle$poll()

## ------------------------------------------------
## Method `crew_class_throttle$new`
## ------------------------------------------------

throttle <- crew_throttle(seconds_interval = 0.5)
throttle$poll()
throttle$poll()
}
\seealso{
Other class: 
\code{\link{crew_class_async}},
\code{\link{crew_class_client}},
\code{\link{crew_class_controller_group}},
\code{\link{crew_class_controller}},
\code{\link{crew_class_launcher}},
\code{\link{crew_class_relay}},
\code{\link{crew_class_tls}}
}
\concept{class}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{seconds_interval}}{Positive numeric of length 1,
throttling interval in seconds.}

\item{\code{polled}}{Positive numeric of length 1,
millisecond timestamp of the last time \code{poll()} returned \code{TRUE}.
\code{NULL} if \code{poll()} was never called on the current object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_throttle-new}{\code{crew_class_throttle$new()}}
\item \href{#method-crew_class_throttle-validate}{\code{crew_class_throttle$validate()}}
\item \href{#method-crew_class_throttle-poll}{\code{crew_class_throttle$poll()}}
\item \href{#method-crew_class_throttle-reset}{\code{crew_class_throttle$reset()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_throttle-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_throttle-new}{}}}
\subsection{Method \code{new()}}{
Throttle constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_throttle$new(seconds_interval = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seconds_interval}}{Throttling interval in seconds.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{R6} object with throttle configuration.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{throttle <- crew_throttle(seconds_interval = 0.5)
throttle$poll()
throttle$poll()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_throttle-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_throttle-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_throttle$validate(test = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{test}}{Logical of length 1, whether to test the throttle configuration
with \code{nanonext::throttle_config()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_throttle-poll"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_throttle-poll}{}}}
\subsection{Method \code{poll()}}{
Poll the throttler.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_throttle$poll()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if \code{poll()} did not return \code{TRUE} in the last
\code{seconds_interval} seconds, \code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_throttle-reset"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_throttle-reset}{}}}
\subsection{Method \code{reset()}}{
Reset the throttle object so the next \code{poll()} returns
\code{TRUE}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_throttle$reset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
}
