% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_tls.R
\name{crew_class_tls}
\alias{crew_class_tls}
\title{\code{R6} TLS class.}
\description{
\code{R6} class for TLS configuration.
}
\details{
See \code{\link[=crew_tls]{crew_tls()}}.
}
\examples{
crew_tls(mode = "automatic")

## ------------------------------------------------
## Method `crew_class_tls$new`
## ------------------------------------------------

crew_tls(mode = "automatic")
}
\seealso{
Other class: 
\code{\link{crew_class_client}},
\code{\link{crew_class_controller_group}},
\code{\link{crew_class_controller}},
\code{\link{crew_class_launcher}},
\code{\link{crew_class_schedule}}
}
\concept{class}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of the \code{\link[=crew_client]{crew_client()}} object paired with this TLS
object. Automatically set in \code{\link[=crew_client]{crew_client()}}.}

\item{\code{mode}}{See \code{\link[=crew_tls]{crew_tls()}}.}

\item{\code{key}}{See \code{\link[=crew_tls]{crew_tls()}}.}

\item{\code{password}}{See \code{\link[=crew_tls]{crew_tls()}}.}

\item{\code{certificates}}{See \code{\link[=crew_tls]{crew_tls()}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_tls-new}{\code{crew_class_tls$new()}}
\item \href{#method-crew_class_tls-validate}{\code{crew_class_tls$validate()}}
\item \href{#method-crew_class_tls-validate_mode_automatic}{\code{crew_class_tls$validate_mode_automatic()}}
\item \href{#method-crew_class_tls-validate_mode_custom}{\code{crew_class_tls$validate_mode_custom()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_tls-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_tls-new}{}}}
\subsection{Method \code{new()}}{
TLS configuration constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_tls$new(
  name = NULL,
  mode = NULL,
  key = NULL,
  password = NULL,
  certificates = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the \code{\link[=crew_tls]{crew_tls()}} object paired with this TLS
object. Automatically set in \code{\link[=crew_tls]{crew_tls()}}.}

\item{\code{mode}}{Argument passed from \code{\link[=crew_tls]{crew_tls()}}.}

\item{\code{key}}{Argument passed from \code{\link[=crew_tls]{crew_tls()}}.}

\item{\code{password}}{Argument passed from \code{\link[=crew_tls]{crew_tls()}}.}

\item{\code{certificates}}{Argument passed from \code{\link[=crew_tls]{crew_tls()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{R6} object with TLS configuration.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crew_tls(mode = "automatic")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_tls-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_tls-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_tls$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_tls-validate_mode_automatic"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_tls-validate_mode_automatic}{}}}
\subsection{Method \code{validate_mode_automatic()}}{
Validation for non-custom modes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_tls$validate_mode_automatic()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_tls-validate_mode_custom"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_tls-validate_mode_custom}{}}}
\subsection{Method \code{validate_mode_custom()}}{
Validation for custom mode.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_tls$validate_mode_custom()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
}
