% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_condition.R
\name{crew_condition_false}
\alias{crew_condition_false}
\title{Return \code{FALSE} on error in a \code{tryCatch()} statement.}
\usage{
crew_condition_false(condition)
}
\arguments{
\item{condition}{Condition object in the \code{tryCatch()} statement.}
}
\value{
\code{FALSE}
}
\description{
For internal use only. Not a user-side function.
Do not invoke directly.
}
\examples{
tryCatch(stop("error message"), error = crew_condition_false)
}
\keyword{internal}
