% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_condition.R
\name{crew_condition_message}
\alias{crew_condition_message}
\title{Return the condition message in a \code{tryCatch()} statement.}
\usage{
crew_condition_message(condition)
}
\arguments{
\item{condition}{Condition object in the \code{tryCatch()} statement.}
}
\value{
Character string with the condition message.
}
\description{
For internal use only. Not a user-side function.
Do not invoke directly.
}
\examples{
tryCatch(stop("error message"), error = crew_condition_message)
}
\keyword{internal}
