% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_controller_group.R
\name{crew_controller_group}
\alias{crew_controller_group}
\title{Create a controller group.}
\usage{
crew_controller_group(...)
}
\arguments{
\item{...}{\code{R6} controller objects or lists of \code{R6} controller objects.
Nested lists are allowed, but each element must be a control object
or another list.}
}
\description{
Create an \code{R6} object to submit tasks and launch workers
through multiple \code{crew} controllers.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_session_start()
persistent <- crew_controller_callr(name = "persistent")
transient <- crew_controller_callr(
  name = "transient",
  max_tasks = 1L
)
group <- crew_controller_group(persistent, transient)
group$connect()
group$push(name = "task", command = sqrt(4), controller = "transient")
group$wait()
group$pop()
group$controllers[["persistent"]]$launcher$running() # 0
group$controllers[["transient"]]$launcher$running() # 0
group$terminate()
crew_session_terminate()
}
}
\seealso{
Other controllers: 
\code{\link{crew_class_controller_group}},
\code{\link{crew_class_controller}},
\code{\link{crew_controller_callr}()},
\code{\link{crew_controller}()}
}
\concept{controllers}
\keyword{internal}
