% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_condition.R
\name{crew_error}
\alias{crew_error}
\title{Throw a crew error.}
\usage{
crew_error(message = NULL, class = character(0))
}
\arguments{
\item{message}{Character of length 1, error message to print.}

\item{class}{Character vector of S3 classes to}
}
\value{
Throw an error of class \code{c("crew_error", "crew")}.
}
\description{
For internal use only. Not a user-side function.
Do not invoke directly.
}
\examples{
try(crew_error("custom error message"))
}
\keyword{internal}
