% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_eval.R
\name{crew_eval}
\alias{crew_eval}
\title{Evaluate an R command and return results as a monad.}
\usage{
crew_eval(command, envir = parent.frame())
}
\arguments{
\item{command}{Language object with R code to run.}

\item{envir}{Environment to run \code{command}.}
}
\value{
A monad object with results and metadata.
}
\description{
Not a user-side function. Do not call directly.
}
\examples{
crew_eval(quote(1 + 1))
}
\seealso{
Other utilities: 
\code{\link{crew_wait}()},
\code{\link{crew_worker}()}
}
\concept{utilities}
\keyword{internal}
