% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_worker.R
\name{crew_log}
\alias{crew_log}
\title{Print a \code{\link[=crew_worker]{crew_worker()}} log message.}
\usage{
crew_log(...)
}
\arguments{
\item{...}{Character strings to print.}
}
\value{
Nothing (invisibly) but print a message.
}
\description{
Not a user-side function. Do not invoke directly.
}
\examples{
crew_log("done")
}
\keyword{internal}
