% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_worker_local.R
\name{crew_loop_worker_local}
\alias{crew_loop_worker_local}
\title{Local worker event loop.}
\usage{
crew_loop_worker_local(name, dir_root, timeout, wait_input)
}
\arguments{
\item{name}{Character of length 1, name of the worker.}

\item{dir_root}{Character of length 1, directory path to the data store.}

\item{timeout}{Positive numeric of length 1, number of seconds
that a worker can idle before timing out.}

\item{wait_input}{Positive numeric of length 1, number of seconds
that the worker waits between checking if a job exists.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Local worker event loop.
}
