% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_mirai_router.R
\name{crew_mirai_router}
\alias{crew_mirai_router}
\title{Create a \code{mirai} router.}
\usage{
crew_mirai_router(host = NULL, port = NULL)
}
\arguments{
\item{host}{IP address the router uses to connect to the workers.
If a character string, the router uses the specified IP address.
If \code{NULL}, the IP address defaults to \code{getip::getip(type = "local")}.}

\item{port}{TCP port the router uses to connect to the workers.
If an integer or character string, the router uses the specified port.
If \code{NULL}, the port defaults to an available port in the ephemeral
port range (49152 through 65355).}
}
\description{
Create an \code{R6} object to route tasks to \code{mirai} workers
on the local network.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_mirai_router()
router$validate()
router$connect()
# Now launch workers with a mirai launcher.
router$disconnect()
}
}
\seealso{
Other routers: 
\code{\link{crew_class_mirai_router}}
}
\concept{routers}
