% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_monitor_local.R
\name{crew_monitor_local}
\alias{crew_monitor_local}
\title{Create a local monitor object.}
\usage{
crew_monitor_local()
}
\description{
Create an \code{R6} object to monitor local processes created by
\code{crew} and \code{mirai}.
}
\seealso{
Other monitor: 
\code{\link{crew_class_monitor_local}}
}
\concept{monitor}
