% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_options_metrics.R
\name{crew_options_metrics}
\alias{crew_options_metrics}
\title{Options for logging resource usage metrics.}
\usage{
crew_options_metrics(
  path_local = NULL,
  path_workers = NULL,
  seconds_local = 5,
  seconds_workers = 5
)
}
\arguments{
\item{path_local}{Where to write resource metric log entries for the
local R session and \code{mirai} dispatcher process.
If \code{NULL}, resource metric logging is turned off for these processes.
If \code{"/dev/stdout"} or \code{"/dev/stderr"}, resource usage metrics
are printed to standard output or standard error,
respectively. If a different character string, that string is the
directory path for writing logs to files on disk, and each
controller instance gets a different log file.
\code{\link[autometric:log_read]{autometric::log_read()}} and \code{\link[autometric:log_plot]{autometric::log_plot()}} can read and
visualize data from logs, even if the logs files are mixed
with other kinds of messages.}

\item{path_workers}{Like \code{path_local}, but for the \code{crew} worker processes.
On Unix-like systems, it is recommended to set \code{path_workers}
equal to \code{"/dev/stdout"} or \code{"/dev/stderr"} to automatically write
resource log messages to the existing log files generated on your
platform (for example, the logs configured with
\code{\link[=crew_options_local]{crew_options_local()}} in the case of \code{\link[=crew_controller_local]{crew_controller_local()}}).
\code{\link[autometric:log_read]{autometric::log_read()}} and \code{\link[autometric:log_plot]{autometric::log_plot()}} can read and
visualize data from logs, even if the logs files are mixed
with other kinds of messages.}

\item{seconds_local}{Positive number, seconds between resource metric log
entries written to \code{path_local}.}

\item{seconds_workers}{Positive number, seconds between resource metric log
entries written to \code{seconds_workers}.}
}
\value{
A classed list of options for logging resource usage metrics.
}
\description{
If a \code{\link[=crew_options_metrics]{crew_options_metrics()}} object is
supplied to the \code{options_metrics} argument of a \code{crew}
controller function, then the \code{autometric} R package will
record resource usage metrics (such as CPU and memory usage)
as the controller and workers are running. Logging happens in
the background (through a detached POSIX) so as not to disrupt
the R session. On Unix-like systems, \code{\link[=crew_options_metrics]{crew_options_metrics()}}
can specify \verb{/dev/stdout} or \verb{/dev/stderr} as the log files, which will
redirect output to existing logs you are already using. Then,
even if those logs are mixed with other messages, you can use functions
\code{\link[autometric:log_read]{autometric::log_read()}} and \code{\link[autometric:log_plot]{autometric::log_plot()}} to read and
visualize resource usage data.
}
\examples{
  crew_options_metrics()
}
\seealso{
Other options: 
\code{\link{crew_options_local}()}
}
\concept{options}
