% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_options_ssh.R
\name{crew_options_ssh}
\alias{crew_options_ssh}
\title{SSH options.}
\usage{
crew_options_ssh(host, port = 22L, seconds_timeout = 10, command = "ssh")
}
\arguments{
\item{host}{A character vector of one or more local IP addresses or
host names of remote machines running \code{crew} workers.
These IPs/hosts must be known and already running before you
start the controller.
The length of \code{host} equals the number of workers that
will launch, and you can repeat the same IP addresses if you want
to launch multiple workers on the same machine.
In fact, \code{host} overrides the \code{workers} argument of the launcher.}

\item{port}{Positive integer of length 1, port number to use for SSH.}

\item{seconds_timeout}{Positive numeric of length 1,
number of seconds to allow before timing out.
See the \code{timeout} argument in \code{?mirai::ssh_config} for details.}

\item{command}{Character string, command used to run SSH.
See the \code{command} argument in \code{?mirai::ssh_config} for details.}
}
\value{
A classed list of options for SSH connections.
}
\description{
\code{\link[=crew_options_ssh]{crew_options_ssh()}} configures the
\code{crew} controller to connect to workers through
SSH connections.
This is useful when the workers run on remote machines and the
local IP addresses or host names of the machines running those workers
are known in advance.
}
\details{
\code{crew} controllers with SSH works differently than the
other controllers.
With SSH, \code{crew} launches all its workers in advance, and it
keeps them running for the duration of the controller.
In addition, the machines you specify with \code{host} have to exist and
already be running in advance.
}
\examples{
  crew_options_ssh(host = "10.0.0.4")
}
\seealso{
Other options: 
\code{\link{crew_options_local}()},
\code{\link{crew_options_metrics}()}
}
\concept{options}
