% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_options.R
\name{crew_options_validate}
\alias{crew_options_validate}
\title{Validate options.}
\usage{
crew_options_validate(options)
}
\value{
\code{NULL} (invisibly) on success, throws an error if there is
something wrong with the options list.
}
\description{
Validate a \code{crew} options list.
}
\concept{utilities}
