% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_port_set.R
\docType{data}
\name{crew_port_envir}
\alias{crew_port_envir}
\title{Set the port for monitoring workers.}
\format{
An object of class \code{environment} of length 0.
}
\usage{
crew_port_envir
}
\description{
Set a TCP port to monitor updates from workers.
}
\keyword{port}
