% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_port_get.R
\name{crew_port_get}
\alias{crew_port_get}
\title{Get the reserved TCP port for monitoring workers.}
\usage{
crew_port_get()
}
\value{
An integer of length 1 with the port number if
the port was set, \code{NULL} if the port has not been set yet.
}
\description{
Get the reserved TCP port for monitoring workers.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_port_set()
crew_port_get()
crew_port_unset()
}
}
\keyword{port}
