% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_port_set.R
\name{crew_port_set}
\alias{crew_port_set}
\title{Reserve a port for monitoring workers.}
\usage{
crew_port_set(port = NULL)
}
\arguments{
\item{port}{Positive integer with the port number. Ephemeral
ports between 49152 and 65535 are recommended.
If \code{port} is \code{NULL}, a port will be chosen from
\code{parallelly::freePort()}.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Reserve a TCP port to monitor updates from workers.
}
\details{
When you set the port, it is reserved with an NNG socket.
\code{\link[=crew_port_unset]{crew_port_unset()}} closes the socket and frees the port.
Be sure to call \code{\link[=crew_port_unset]{crew_port_unset()}} when you are finished using
\code{crew} in your R session.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_port_set()
crew_port_get()
crew_port_unset()
}
}
\keyword{port}
