% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_port_unset.R
\name{crew_port_unset}
\alias{crew_port_unset}
\title{Unset and release the TCP port for monitoring workers.}
\usage{
crew_port_unset()
}
\value{
\code{NULL} (invisibly).
}
\description{
Unset and release the TCP port for monitoring workers.
}
\details{
This function closes the NNG socket that reserved the port,
freeing it for other programs to use. If you previously set a port with
\code{\link[=crew_port_set]{crew_port_set()}}, always call \code{crew_port_unset()} when you
are done using \code{crew} for the current R session.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_port_set()
crew_port_get()
crew_port_unset()
}
}
\keyword{port}
