% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_redis_server.R
\name{crew_redis_server}
\alias{crew_redis_server}
\title{Redis server}
\usage{
crew_redis_server(
  binary = NULL,
  conf = NULL,
  host = NULL,
  port = NULL,
  password = NULL,
  start_timeout = 5,
  start_wait = 0.25
)
}
\arguments{
\item{binary}{Nonempty non-missing character of length 1,
path to the Redis server executable.
The path must point to an existing file.
If the \code{binary} argument is \code{NULL},
then the value defaults to \code{Sys.getenv("CREW_REDIS_SERVER_BINARY")}
if this environment variable is set. Otherwise, the default becomes
\code{unname(Sys.which("redis-server"))}.
Visit \url{https://redis.io/} to learn how to install Redis server.}

\item{conf}{Non-missing character of length 1,
path to the Redis server configuration file.
The default value of the \code{conf} argument
is \code{Sys.getenv("CREW_REDIS_SERVER_CONF")} if this environment
variable is set. Otherwise, the default becomes
\code{""} to use the system default configuration file.
Visit \url{https://redis.io/docs/management/config/}
to learn how to configure Redis.}

\item{host}{Nonempty non-missing character of length 1,
IP address of the Redis server.
The default value is \code{Sys.getenv("CREW_REDIS_SERVER_HOST")},
if this environment variable is set. Otherwise, the default value
is the IPv4 loopback address \code{"127.0.0.1"}.}

\item{port}{Nonempty non-missing character of length 1,
TCP port of the Redis server.
The default value is \code{Sys.getenv("CREW_REDIS_SERVER_PORT")}
if this environment variable is set. Otherwise, the default is
a random ephemeral port.}

\item{password}{Non-missing character of length 1
with at most 64 characters,
password to the Redis server (default user).
The default value is \code{Sys.getenv("CREW_REDIS_SERVER_PASSWORD")}
if this environment variable is set. Otherwise, the default is a
a random cryptographic 256-bit 64-character password.
Visit \url{https://redis.io/docs/management/security/} to learn about
Redis security.}

\item{start_timeout}{Positive numeric of length 1, number of seconds
to wait for the Redis server to start up and begin accepting clients.}

\item{start_wait}{Positive numeric of length 1, polling interval
(seconds) while waiting for the Redis server to start up and
begin accepting clients.}
}
\value{
An \code{R6} object with methods to configure, run,
and manage an instance of Redis server.
}
\description{
Create a handle to configure, run, and manage
an instance of Redis server.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
server <- crew_redis_server()
server$validate()
server$test() # TRUE
server$start()
server$ping() # [Redis: PONG]
server$stop()
}
}
\concept{redis}
