% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_relay.R
\name{crew_relay}
\alias{crew_relay}
\title{Create a \code{crew} relay object.}
\usage{
crew_relay(seconds_interval = 1)
}
\arguments{
\item{seconds_interval}{Positive numeric scalar, maximum wait time interval
for throttling (see \code{\link[=crew_throttle]{crew_throttle()}}).}
}
\value{
An \code{R6} \code{crew} relay object.
}
\description{
Create an \code{R6} \code{crew} relay object.
}
\details{
A \code{crew} relay object keeps the signaling relationships
among condition variables.
}
\examples{
crew_relay()
}
\seealso{
Other relay: 
\code{\link{crew_class_relay}}
}
\concept{relay}
