% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_schedule.R
\name{crew_schedule}
\alias{crew_schedule}
\title{Create a schedule.}
\usage{
crew_schedule(seconds_interval = 0.25)
}
\arguments{
\item{seconds_interval}{Number of seconds between throttled iterations
of task collection.}
}
\description{
Create an \code{R6} object to contain and manage task objects.
}
\details{
Not a user-side function. There are no examples. Please see
\code{\link[=crew_controller_local]{crew_controller_local()}} for details.
}
\seealso{
Other schedule: 
\code{\link{crew_class_schedule}}
}
\concept{schedule}
\keyword{internal}
