% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_session_close.R
\name{crew_session_close}
\alias{crew_session_close}
\title{Close the crew session.}
\usage{
crew_session_close()
}
\value{
\code{NULL} (invisibly).
}
\description{
Close the crew session.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_session_open()
crew_session_port()
crew_session_close()
}
}
\keyword{session}
