% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_session_open.R
\name{crew_session_open}
\alias{crew_session_open}
\title{Open a new crew session.}
\usage{
crew_session_open(host = NULL, port = NULL)
}
\arguments{
\item{host}{IP address of the machine hosting the session.
Defaults to the local IP address of the session.}

\item{port}{Positive integer with the port number. Ephemeral
ports between 49152 and 65535 are recommended.
If \code{port} is \code{NULL}, a port will be chosen from
\code{parallelly::freePort()}.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Open a crew session if a session does not already exist.
}
\details{
A \code{crew} session reserves a TCP port with a new NNG socket.
\code{crew} uses this port to monitor the connection statuses of
all workers launched from the current R session across all
controllers and multi-controllers.
\code{\link[=crew_session_close]{crew_session_close()}} closes the NNG socket and frees the port.
Be sure to call \code{\link[=crew_session_close]{crew_session_close()}} when you are finished using
\code{crew} in your R session.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_port_set()
crew_port_get()
crew_port_unset()
}
}
\keyword{session}
