% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_session_close.R,
%   R/crew_session_terminate.R
\name{crew_session_terminate}
\alias{crew_session_terminate}
\title{Terminate the crew session.}
\usage{
crew_session_terminate()

crew_session_terminate()
}
\value{
\code{NULL} (invisibly).

\code{NULL} (invisibly).
}
\description{
Terminate the crew session.

Close the crew session.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_session_start()
crew_session_port()
crew_session_terminate()
}
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_session_start()
crew_session_port()
crew_session_terminate()
}
}
\keyword{session}
