% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_sitrep_redis.R
\name{crew_sitrep_redis}
\alias{crew_sitrep_redis}
\title{Check Redis.}
\usage{
crew_sitrep_redis()
}
\value{
Nothing (invisibly). This function prints out messages
to describe the status of the Redis installation and
configuration.
}
\description{
Check the Redis installation and configuration
and print messages with any obvious issues.
}
\examples{
crew_sitrep_redis()
}
\concept{sitrep}
