% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_stage.R
\name{crew_stage}
\alias{crew_stage}
\title{Create a \code{crew} stage object.}
\usage{
crew_stage()
}
\value{
An \code{R6} \code{crew} stage object.
}
\description{
Create an \code{R6} \code{crew} stage object.
}
\details{
A \code{crew} stage object keeps track of tasks that are resolved
but not yet popped.
}
\examples{
crew_stage()
}
\seealso{
Other developer: 
\code{\link{crew_async}()},
\code{\link{crew_client}()},
\code{\link{crew_controller}()},
\code{\link{crew_launcher}()}
}
\concept{developer}
