% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_store_local.R
\name{crew_store_local}
\alias{crew_store_local}
\title{Local store}
\description{
Local store class.
}
\details{
Most \code{crew} queues use data stores to send and receive task data.
Some use entirely local data, others may use cloud data to support
cloud workers. The local store works with files on the local file system.
}
\examples{
# Usage on its own:
store <- crew_store_local$new(timeout = Inf)
store$write_main_input("worker", list(value = "x"))
store$list_main_input()
store$exists_main_input("worker")
store$read_main_input("worker")
# Usage in a queue:
fun <- function(x) x + 1
args <- list(x = 1)
queue <- crew_queue_future$new(timeout = 60, store = store)
queue$push(fun = fun, args = args)
queue$block()
result <- queue$pop()
str(result)
result$result$result
queue$shutdown()
processx::supervisor_kill()
store$destroy()
}
\seealso{
Other store: 
\code{\link{crew_store}}
}
\concept{store}
\section{Super class}{
\code{\link[crew:crew_store]{crew::crew_store}} -> \code{crew_store_local}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-read_worker_input}{\code{crew_store_local$read_worker_input()}}
\item \href{#method-read_worker_output}{\code{crew_store_local$read_worker_output()}}
\item \href{#method-write_worker_input}{\code{crew_store_local$write_worker_input()}}
\item \href{#method-write_worker_output}{\code{crew_store_local$write_worker_output()}}
\item \href{#method-exists_worker_input}{\code{crew_store_local$exists_worker_input()}}
\item \href{#method-exists_worker_output}{\code{crew_store_local$exists_worker_output()}}
\item \href{#method-list_worker_input}{\code{crew_store_local$list_worker_input()}}
\item \href{#method-list_worker_output}{\code{crew_store_local$list_worker_output()}}
\item \href{#method-delete_worker_input}{\code{crew_store_local$delete_worker_input()}}
\item \href{#method-delete_worker_output}{\code{crew_store_local$delete_worker_output()}}
\item \href{#method-upload_input}{\code{crew_store_local$upload_input()}}
\item \href{#method-download_output}{\code{crew_store_local$download_output()}}
\item \href{#method-marshal}{\code{crew_store_local$marshal()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="crew_store" data-id="delete_main_input">}\href{../../crew/html/crew_store.html#method-delete_main_input}{\code{crew::crew_store$delete_main_input()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="crew_store" data-id="delete_main_output">}\href{../../crew/html/crew_store.html#method-delete_main_output}{\code{crew::crew_store$delete_main_output()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="crew_store" data-id="destroy">}\href{../../crew/html/crew_store.html#method-destroy}{\code{crew::crew_store$destroy()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="crew_store" data-id="exists_main_input">}\href{../../crew/html/crew_store.html#method-exists_main_input}{\code{crew::crew_store$exists_main_input()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="crew_store" data-id="exists_main_output">}\href{../../crew/html/crew_store.html#method-exists_main_output}{\code{crew::crew_store$exists_main_output()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="crew_store" data-id="get_root">}\href{../../crew/html/crew_store.html#method-get_root}{\code{crew::crew_store$get_root()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="crew_store" data-id="initialize">}\href{../../crew/html/crew_store.html#method-initialize}{\code{crew::crew_store$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="crew_store" data-id="list_main_input">}\href{../../crew/html/crew_store.html#method-list_main_input}{\code{crew::crew_store$list_main_input()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="crew_store" data-id="list_main_output">}\href{../../crew/html/crew_store.html#method-list_main_output}{\code{crew::crew_store$list_main_output()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="crew_store" data-id="read_main_input">}\href{../../crew/html/crew_store.html#method-read_main_input}{\code{crew::crew_store$read_main_input()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="crew_store" data-id="read_main_output">}\href{../../crew/html/crew_store.html#method-read_main_output}{\code{crew::crew_store$read_main_output()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="crew_store" data-id="write_main_input">}\href{../../crew/html/crew_store.html#method-write_main_input}{\code{crew::crew_store$write_main_input()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="crew_store" data-id="write_main_output">}\href{../../crew/html/crew_store.html#method-write_main_output}{\code{crew::crew_store$write_main_output()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_worker_input"></a>}}
\if{latex}{\out{\hypertarget{method-read_worker_input}{}}}
\subsection{Method \code{read_worker_input()}}{
Read worker input data from the worker file space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_store_local$read_worker_input(worker)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{worker}}{Character of length 1, name of the worker.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Worker input data from the worker file space.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_worker_output"></a>}}
\if{latex}{\out{\hypertarget{method-read_worker_output}{}}}
\subsection{Method \code{read_worker_output()}}{
Read worker output data from the worker file space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_store_local$read_worker_output(worker)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{worker}}{Character of length 1, name of the worker.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Worker output data from the worker file space.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_worker_input"></a>}}
\if{latex}{\out{\hypertarget{method-write_worker_input}{}}}
\subsection{Method \code{write_worker_input()}}{
Write worker input data to the worker file space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_store_local$write_worker_input(worker, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{worker}}{Character of length 1, name of the worker.}

\item{\code{value}}{R object to write.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_worker_output"></a>}}
\if{latex}{\out{\hypertarget{method-write_worker_output}{}}}
\subsection{Method \code{write_worker_output()}}{
Write worker output data to the worker file space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_store_local$write_worker_output(worker, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{worker}}{Character of length 1, name of the worker.}

\item{\code{value}}{R object to output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exists_worker_input"></a>}}
\if{latex}{\out{\hypertarget{method-exists_worker_input}{}}}
\subsection{Method \code{exists_worker_input()}}{
Check if worker input data exists in the worker file space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_store_local$exists_worker_input(worker)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{worker}}{Character of length 1, name of the worker.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical of length 1, whether worker input data exists
in the worker file space.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exists_worker_output"></a>}}
\if{latex}{\out{\hypertarget{method-exists_worker_output}{}}}
\subsection{Method \code{exists_worker_output()}}{
Check if worker output data exists in the worker file space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_store_local$exists_worker_output(worker)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{worker}}{Character of length 1, name of the worker.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical of length 1, whether worker output data exists
in the worker file space.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_worker_input"></a>}}
\if{latex}{\out{\hypertarget{method-list_worker_input}{}}}
\subsection{Method \code{list_worker_input()}}{
List workers with input data in the worker file space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_store_local$list_worker_input()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of workers with input data
in the worker file space.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_worker_output"></a>}}
\if{latex}{\out{\hypertarget{method-list_worker_output}{}}}
\subsection{Method \code{list_worker_output()}}{
List workers with output data in the worker file space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_store_local$list_worker_output()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of workers with output data
in the worker file space.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_worker_input"></a>}}
\if{latex}{\out{\hypertarget{method-delete_worker_input}{}}}
\subsection{Method \code{delete_worker_input()}}{
Delete worker input data from the worker file space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_store_local$delete_worker_input(worker)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{worker}}{Character of length 1, name of the worker.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_worker_output"></a>}}
\if{latex}{\out{\hypertarget{method-delete_worker_output}{}}}
\subsection{Method \code{delete_worker_output()}}{
Delete worker output data from the worker file space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_store_local$delete_worker_output(worker)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{worker}}{Character of length 1, name of the worker.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload_input"></a>}}
\if{latex}{\out{\hypertarget{method-upload_input}{}}}
\subsection{Method \code{upload_input()}}{
Upload worker input data from the main file space
to the worker file space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_store_local$upload_input(worker)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{worker}}{Character of length 1, name of the worker.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
For the local store, this is not a true upload. It is
just moving a file.
}

\subsection{Returns}{
\code{NULL} (invisibly)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_output"></a>}}
\if{latex}{\out{\hypertarget{method-download_output}{}}}
\subsection{Method \code{download_output()}}{
Download worker input data from the worker file space
to the main file space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_store_local$download_output(worker)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{worker}}{Character of length 1, name of the worker.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
For the local store, this is not a true upload. It is
just moving a file.
}

\subsection{Returns}{
\code{NULL} (invisibly)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-marshal"></a>}}
\if{latex}{\out{\hypertarget{method-marshal}{}}}
\subsection{Method \code{marshal()}}{
Marshal the data store.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_store_local$marshal()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Represent the store object as a concise
character string that can be recovered
with \code{parse()} and \code{eval()} in a different R process.
}

\subsection{Returns}{
A concise character string that returns the data store
object when evaluated with \code{parse()} and \code{eval()}
in a different R process.
}
}
}
