% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_terminate_signal.R
\name{crew_terminate_signal}
\alias{crew_terminate_signal}
\title{Get the termination signal.}
\usage{
crew_terminate_signal()
}
\value{
An integer of length 1: \code{tools::SIGTERM} if your platform
supports \code{SIGTERM}. If not, then \code{crew_crew_terminate_signal()()} checks
\code{SIGQUIT}, then \code{SIGINT}, then \code{SIGKILL}, and then returns the first
signal it finds that your operating system can use.
}
\description{
Get a supported operating system signal for terminating
a local process.
}
\examples{
crew_terminate_signal()
}
\seealso{
Other utility: 
\code{\link{crew_assert}()},
\code{\link{crew_clean}()},
\code{\link{crew_deprecate}()},
\code{\link{crew_eval}()},
\code{\link{crew_random_name}()},
\code{\link{crew_retry}()},
\code{\link{crew_terminate_process}()},
\code{\link{crew_worker}()}
}
\concept{utility}
