% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_throttle.R
\name{crew_throttle}
\alias{crew_throttle}
\title{Create a stateful throttling object.}
\usage{
crew_throttle(seconds_interval = 0.5)
}
\arguments{
\item{seconds_interval}{Positive numeric of length 1, throttling interval.
The \code{poll()} method returns \code{TRUE} if and only if it was not called
in the last \code{seconds_interval} seconds.}
}
\value{
An \code{R6} object with throttle configuration settings and methods.
}
\description{
Create an \code{R6} object for throttling.
}
\details{
Throttling is a technique that limits how often a function is
called in a given period of time. You can use it to help \code{crew}
be kinder to the \code{mirai} dispatcher. For example,
if you plan to push a lot of tasks and want to avoid the burden
of so many scaling operations, set
\code{throttle <- crew_throttle(seconds_interval = 0.5)} and then call
\verb{controller$push(..., scale = throttle$poll()} (where \code{...} is
shorthand for the other arguments you supply to \code{push()}).
\code{poll()} returns \code{TRUE} if and only if it did not return \code{TRUE}
in the last \code{seconds_interval} seconds,
so the overall effect is to scale every
0.5 seconds even if \code{push()} is called more frequently than that.
}
\examples{
throttle <- crew_throttle(seconds_interval = 0.5)
throttle$poll()
throttle$poll()
}
\seealso{
Other user: 
\code{\link{crew-package}},
\code{\link{crew_clean}()},
\code{\link{crew_controller_group}()},
\code{\link{crew_controller_local}()},
\code{\link{crew_tls}()}
}
\concept{user}
