% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_condition.R
\name{crew_true}
\alias{crew_true}
\title{Validate arguments}
\usage{
crew_true(value = NULL, ..., message = NULL, envir = parent.frame())
}
\arguments{
\item{value}{Data object to validate.}

\item{...}{Assertions to check.}

\item{message}{Optional character of length 1 with a message to print.}

\item{envir}{Optional environment to evaluate the assertions.}
}
\value{
Nothing (invisibly) if the object is successfully validated,
throws an error of class \code{"crew_error"}
if at least one of the assertions is not true.
}
\description{
Validate input arguments of \code{crew} functions.
}
\details{
Used internally and in packages that extend \code{crew}.
}
\examples{
crew_true(letters, length(.) > 0, is.character(.), !anyNA(.))
try(crew_true(letters, is.numeric(.), !anyNA(.)))
}
\seealso{
Other utilities: 
\code{\link{crew_wait}()}
}
\concept{utilities}
