% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_user.R
\name{crew_user}
\alias{crew_user}
\title{Get the user ID.}
\usage{
crew_user()
}
\value{
Character of length 1 with the user ID if it can be determined,
\code{NULL} if the user ID cannot be determined.
}
\description{
Get the user ID running the current R session.
}
\details{
First try \code{Sys.info()["user"]}. If the value is \code{"unknown"},
try \code{Sys.info()["effective_user"]}. If that values is also
\code{"unknown"}, then the return value is \code{NULL}.
}
\seealso{
Other utility: 
\code{\link{crew_assert}()},
\code{\link{crew_clean}()},
\code{\link{crew_deprecate}()},
\code{\link{crew_eval}()},
\code{\link{crew_random_name}()},
\code{\link{crew_retry}()},
\code{\link{crew_signal_terminate}()},
\code{\link{crew_worker}()}
}
\concept{utility}
