% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_crew.R
\name{class_crew}
\alias{class_crew}
\alias{crew}
\title{Crew class.}
\description{
\code{R6} class for a crew.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Character of length 1, crew name.}

\item{\code{store}}{\code{R6} store object.}

\item{\code{worker_classes}}{Named list of \code{R6ClassGenerator}
objects with worker definitions. The first one
is the default.}

\item{\code{workers}}{Named list of worker objects.}

\item{\code{max_workers}}{Positive integer,
maximum number of workers in the crew.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{class_crew$new()}}
\item \href{#method-recruit}{\code{class_crew$recruit()}}
\item \href{#method-launch}{\code{class_crew$launch()}}
\item \href{#method-sendable}{\code{class_crew$sendable()}}
\item \href{#method-send}{\code{class_crew$send()}}
\item \href{#method-receivable}{\code{class_crew$receivable()}}
\item \href{#method-receive}{\code{class_crew$receive()}}
\item \href{#method-shutdown}{\code{class_crew$shutdown()}}
\item \href{#method-dismiss}{\code{class_crew$dismiss()}}
\item \href{#method-validate}{\code{class_crew$validate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Crew constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_crew$new(
  name = basename(tempfile(pattern = "crew_")),
  store = crew::class_store_local$new(),
  worker_classes = list(crew::class_worker_callr, crew::class_worker_future),
  max_workers = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character of length 1, crew name.}

\item{\code{store}}{\code{R6} store object.}

\item{\code{worker_classes}}{List of \code{R6ClassGenerator}
objects (created with \code{R6Class}) with worker definitions
and constructor.}

\item{\code{max_workers}}{Positive integer,
maximum number of workers in the crew.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recruit"></a>}}
\if{latex}{\out{\hypertarget{method-recruit}{}}}
\subsection{Method \code{recruit()}}{
create worker objects from one or more worker definitions.
Does not actually launch the new worker objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_crew$recruit(workers = 1L, class = names(self$worker_classes)[[1]], ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workers}}{Number of new worker objects to create.}

\item{\code{class}}{Name of the worker class to use for launching workers.
check the \code{worker_classes} field for possible names.}

\item{\code{...}}{Named arguments to the worker constructor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-launch"></a>}}
\if{latex}{\out{\hypertarget{method-launch}{}}}
\subsection{Method \code{launch()}}{
Launch all down workers in the crew.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_crew$launch()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sendable"></a>}}
\if{latex}{\out{\hypertarget{method-sendable}{}}}
\subsection{Method \code{sendable()}}{
Determine if any worker is unassigned
and ready to accept a new job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_crew$sendable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send"></a>}}
\if{latex}{\out{\hypertarget{method-send}{}}}
\subsection{Method \code{send()}}{
Send a job to an available worker. Assumes \code{sendable()}
on the crew returns \code{TRUE}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_crew$send(fun, args = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{Function to run in the job.}

\item{\code{args}}{Named list of arguments to \code{fun()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-receivable"></a>}}
\if{latex}{\out{\hypertarget{method-receivable}{}}}
\subsection{Method \code{receivable()}}{
Determine if any worker in the crew is done
with its current job and the job output is available for collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_crew$receivable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-receive"></a>}}
\if{latex}{\out{\hypertarget{method-receive}{}}}
\subsection{Method \code{receive()}}{
Choose a receivable worker and collect its job output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_crew$receive()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shutdown"></a>}}
\if{latex}{\out{\hypertarget{method-shutdown}{}}}
\subsection{Method \code{shutdown()}}{
Shut down one or more running workers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_crew$shutdown(workers = Inf, sendable_only = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workers}}{Positive integer of length 1. Maximum number of workers
to try to shut down. Does not count running or non-sendable (busy)
workers.}

\item{\code{sendable_only}}{Logical of length 1, whether to skip to another
worker to shut down if the current worker is not sendable.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method loops through the workers
from first to last in the worker list. If the worker
is up (running) and sendable (able to receive a job)
then the worker is sent a shutdown command. Otherwise,
the loop moves on to another worker.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dismiss"></a>}}
\if{latex}{\out{\hypertarget{method-dismiss}{}}}
\subsection{Method \code{dismiss()}}{
delete all worker objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_crew$dismiss(workers = Inf, sendable_only = TRUE, down_only = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workers}}{Positive integer of length 1. Maximum number of workers
to try to dismiss. Does not count down or non-sendable workers
(workers with unfinished jobs).}

\item{\code{sendable_only}}{Logical of length 1, whether to skip to another
worker to dismiss if the current worker is not sendable.}

\item{\code{down_only}}{Logical of length 1, whether to skip to another
worker to dismiss if the current worker is not down.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method loops through the workers
from first to last in the worker list. If the worker
is down and sendable (able to receive a job)
then the worker object is deleted from the crew. Otherwise,
the loop moves on to another worker.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Crew validator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_crew$validate()}\if{html}{\out{</div>}}
}

}
}
