% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_store.R
\name{class_store}
\alias{class_store}
\title{Store class.}
\description{
\code{R6} class for a store.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dir}}{Character of length 1, root file path or prefix
where the files are located.}

\item{\code{dir_workers}}{Character of length 1, worker storage directory.}

\item{\code{dir_temp}}{Character of length 1, directory for temporary files.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{class_store$new()}}
\item \href{#method-validate}{\code{class_store$validate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Store constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_store$new(dir = tempfile())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{Character of length 1, file path or prefix where the files
are located.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Store validator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_store$validate()}\if{html}{\out{</div>}}
}

}
}
