% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_worker.R
\name{class_worker}
\alias{class_worker}
\title{Worker class.}
\description{
\code{R6} class for a worker.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Character of length 1, worker name.}

\item{\code{crew}}{\code{R6} crew object to which the worker belongs.}

\item{\code{timeout}}{Positive numeric of length 1,
number of seconds of idling for the worker to time out.}

\item{\code{wait_input}}{Positive numeric of length 1,
number of seconds for a worker to wait between iterations of
polling input.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{class_worker$new()}}
\item \href{#method-validate}{\code{class_worker$validate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Worker constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker$new(
  name = basename(tempfile(pattern = "worker_")),
  crew = NULL,
  timeout = 60,
  wait_input = 0.1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character of length 1, worker name.}

\item{\code{crew}}{\code{R6} crew object to which the worker belongs.}

\item{\code{timeout}}{Positive numeric of length 1, number of seconds
that a worker can idle before timing out.}

\item{\code{wait_input}}{Positive numeric of length 1, number of seconds
that the worker waits between checking if a job exists.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Worker validator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker$validate()}\if{html}{\out{</div>}}
}

}
}
