% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_worker.R
\name{class_worker}
\alias{class_worker}
\alias{worker}
\title{Worker class.}
\description{
\code{R6} class for a worker.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Character of length 1, worker name.}

\item{\code{crew}}{\code{R6} crew object to which the worker belongs.}

\item{\code{timeout}}{Positive numeric of length 1,
number of seconds of idling for the worker to time out.}

\item{\code{wait_input}}{Positive numeric of length 1,
number of seconds for a worker to wait between iterations of
polling input.}

\item{\code{assigned}}{Logical of length 1, whether the worker
already has a job to do.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{class_worker$new()}}
\item \href{#method-send}{\code{class_worker$send()}}
\item \href{#method-done}{\code{class_worker$done()}}
\item \href{#method-receive}{\code{class_worker$receive()}}
\item \href{#method-shutdown}{\code{class_worker$shutdown()}}
\item \href{#method-validate}{\code{class_worker$validate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Worker constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker$new(
  name = basename(tempfile(pattern = "worker_")),
  crew = NULL,
  timeout = 60,
  wait_input = 0.1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character of length 1, worker name.}

\item{\code{crew}}{\code{R6} crew object to which the worker belongs.}

\item{\code{timeout}}{Positive numeric of length 1, number of seconds
that a worker can idle before timing out.}

\item{\code{wait_input}}{Positive numeric of length 1, number of seconds
that the worker waits between checking if a job exists.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send"></a>}}
\if{latex}{\out{\hypertarget{method-send}{}}}
\subsection{Method \code{send()}}{
Send a job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker$send(fun, args = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{Function to run in the job. Should be completely
self-contained in the body and arguments, without relying
on the closure or global variables in the environment.}

\item{\code{args}}{Named list of arguments to \code{fun}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-done"></a>}}
\if{latex}{\out{\hypertarget{method-done}{}}}
\subsection{Method \code{done()}}{
\code{TRUE} if a worker is done with a job and the
main process can receive the output of the job. \code{FALSE} otherwise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker$done()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-receive"></a>}}
\if{latex}{\out{\hypertarget{method-receive}{}}}
\subsection{Method \code{receive()}}{
Collect the results of a job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker$receive()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shutdown"></a>}}
\if{latex}{\out{\hypertarget{method-shutdown}{}}}
\subsection{Method \code{shutdown()}}{
Gracefully shut down the worker.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker$shutdown()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Worker validator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker$validate()}\if{html}{\out{</div>}}
}

}
}
