% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_worker_future.R
\name{class_worker_future}
\alias{class_worker_future}
\alias{worker}
\title{Future worker class.}
\description{
\code{R6} class for a \code{future} worker.
}
\section{Super class}{
\code{\link[crew:worker]{crew::worker}} -> \code{worker_future}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{future}}{A \code{future::future()} object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-globals}{\code{class_worker_future$globals()}}
\item \href{#method-send}{\code{class_worker_future$send()}}
\item \href{#method-alive}{\code{class_worker_future$alive()}}
\item \href{#method-done}{\code{class_worker_future$done()}}
\item \href{#method-receive}{\code{class_worker_future$receive()}}
\item \href{#method-terminate}{\code{class_worker_future$terminate()}}
\item \href{#method-validate}{\code{class_worker_future$validate()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="initialize">}\href{../../crew/html/worker.html#method-initialize}{\code{crew::worker$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-globals"></a>}}
\if{latex}{\out{\hypertarget{method-globals}{}}}
\subsection{Method \code{globals()}}{
Create the \code{globals} argument of \code{future::future()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker_future$globals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send"></a>}}
\if{latex}{\out{\hypertarget{method-send}{}}}
\subsection{Method \code{send()}}{
Send a job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker_future$send(fun, args = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{Function to run in the job. Should be completely
self-contained in the body and arguments, without relying
on the closure or global variables in the environment.}

\item{\code{args}}{Named list of arguments to \code{fun}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-alive"></a>}}
\if{latex}{\out{\hypertarget{method-alive}{}}}
\subsection{Method \code{alive()}}{
\code{TRUE} if the worker is alive and \code{FALSE} otherwise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker_future$alive()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-done"></a>}}
\if{latex}{\out{\hypertarget{method-done}{}}}
\subsection{Method \code{done()}}{
\code{TRUE} if a worker is done with a job and the
main process can receive the output of the job. \code{FALSE} otherwise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker_future$done()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-receive"></a>}}
\if{latex}{\out{\hypertarget{method-receive}{}}}
\subsection{Method \code{receive()}}{
Collect the results of a job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker_future$receive()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-terminate"></a>}}
\if{latex}{\out{\hypertarget{method-terminate}{}}}
\subsection{Method \code{terminate()}}{
Terminate the worker.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker_future$terminate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Worker validator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker_future$validate()}\if{html}{\out{</div>}}
}

}
}
