% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_worker_future.R
\name{class_worker_future}
\alias{class_worker_future}
\alias{worker_future}
\title{Future worker class.}
\description{
\code{R6} class for a \code{future} worker.
}
\section{Super class}{
\code{\link[crew:worker]{crew::worker}} -> \code{worker_future}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{future}}{A \code{future::future()} object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-up}{\code{class_worker_future$up()}}
\item \href{#method-validate}{\code{class_worker_future$validate()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="done">}\href{../../crew/html/worker.html#method-done}{\code{crew::worker$done()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="initialize">}\href{../../crew/html/worker.html#method-initialize}{\code{crew::worker$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="receive">}\href{../../crew/html/worker.html#method-receive}{\code{crew::worker$receive()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="send">}\href{../../crew/html/worker.html#method-send}{\code{crew::worker$send()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="shutdown">}\href{../../crew/html/worker.html#method-shutdown}{\code{crew::worker$shutdown()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="submit">}\href{../../crew/html/worker.html#method-submit}{\code{crew::worker$submit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-up"></a>}}
\if{latex}{\out{\hypertarget{method-up}{}}}
\subsection{Method \code{up()}}{
\code{TRUE} if the worker is running and \code{FALSE} otherwise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker_future$up()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Worker validator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker_future$validate()}\if{html}{\out{</div>}}
}

}
}
