% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_worker_future_local.R
\name{class_worker_future_local}
\alias{class_worker_future_local}
\alias{worker_future_local}
\title{Local \code{future} worker class.}
\description{
\code{R6} class for a local \code{future} worker.
It is "local" in the sense that the data store
is a local file system. Workers can use \code{future.batchtools}.
}
\section{Super classes}{
\code{\link[crew:worker]{crew::worker}} -> \code{\link[crew:worker_future]{crew::worker_future}} -> \code{worker_future}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-launch}{\code{class_worker_future_local$launch()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="done">}\href{../../crew/html/worker.html#method-done}{\code{crew::worker$done()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="initialize">}\href{../../crew/html/worker.html#method-initialize}{\code{crew::worker$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="receive">}\href{../../crew/html/worker.html#method-receive}{\code{crew::worker$receive()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="send">}\href{../../crew/html/worker.html#method-send}{\code{crew::worker$send()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker" data-id="shutdown">}\href{../../crew/html/worker.html#method-shutdown}{\code{crew::worker$shutdown()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker_future" data-id="alive">}\href{../../crew/html/worker_future.html#method-alive}{\code{crew::worker_future$alive()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="crew" data-topic="worker_future" data-id="validate">}\href{../../crew/html/worker_future.html#method-validate}{\code{crew::worker_future$validate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-launch"></a>}}
\if{latex}{\out{\hypertarget{method-launch}{}}}
\subsection{Method \code{launch()}}{
Launch the worker.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{class_worker_future_local$launch()}\if{html}{\out{</div>}}
}

}
}
