% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_package.R
\docType{package}
\name{crew-package}
\alias{crew-package}
\title{crew: abstract interface specification for semi-persistent
high-performance computing workers}
\description{
The \code{R6} classes of \code{crew} establish a standardized
user interface to high-performance computing technologies.
Unlike its closely related \code{future} package, \code{crew} prioritizes
centralized scheduling, heterogeneous semi-persistent workers,
and user-driven customization. The primary goal is to help
pipeline tools such as such \code{targets} efficiently orchestrate
tasks without having to support individual low-level interfaces
to specific high-performance computing platforms or cloud services.
}
\concept{help}
