% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_package.R
\docType{package}
\name{crew-package}
\alias{crew-package}
\title{crew: a distributed worker launcher framework}
\description{
In computationally demanding analysis projects,
statisticians and data scientists asynchronously deploy
long-running tasks to distributed systems, ranging from
traditional clusters to cloud services.
The \href{https://nng.nanomsg.org}{NNG}-powered
\href{https://github.com/shikokuchuo/mirai}{\code{mirai}}
R package is a sleek and sophisticated scheduler
that efficiently processes these intense workloads.
The \code{crew} package extends
\href{https://github.com/shikokuchuo/mirai}{\code{mirai}}
with a unifying interface
for third-party worker launchers.
Inspiration also comes from packages
\href{https://future.futureverse.org/}{\code{future}},
\href{https://mrc-ide.github.io/rrq/}{\code{rrq}},
\href{https://mschubert.github.io/clustermq/}{\code{clustermq}},
and \href{https://mllg.github.io/batchtools/}{\code{batchtools}}.
}
\seealso{
Other user: 
\code{\link{crew_controller_group}()},
\code{\link{crew_controller_local}()}
}
\concept{user}
