% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_launcher.R
\name{crew_class_launcher}
\alias{crew_class_launcher}
\title{Launcher abstract class}
\description{
\code{R6} abstract class to build other subclasses
which launch and manage workers.
}
\seealso{
Other launchers: 
\code{\link{crew_class_launcher_callr}},
\code{\link{crew_launcher_callr}()}
}
\concept{launchers}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{idle_time}}{Maximum number of seconds that a worker can idle
since the completion of the last task. If exceeded, the worker exits.}

\item{\code{wall_time}}{Soft wall time in seconds. See the \code{wall_time}
argument of \code{mirai::server()}.}

\item{\code{poll_high}}{High polling interval in seconds for the \code{mirai}
active queue.}

\item{\code{poll_low}}{Low polling interval in seconds for the \code{mirai}
active queue.}

\item{\code{launch_timeout}}{Number of seconds to time out
waiting for a new group of workers to launch.}

\item{\code{launch_wait}}{Number of seconds to wait between checks
that newly launched workers are ready to receive tasks.}

\item{\code{max_tasks}}{Maximum number of tasks that a worker will do before
exiting.}

\item{\code{async_dial}}{Logical, whether the \code{mirai} workers should dial in
asynchronously. See the \code{asyncdial} argument of \code{mirai::server()}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_launcher-new}{\code{crew_class_launcher$new()}}
\item \href{#method-crew_class_launcher-validate}{\code{crew_class_launcher$validate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher-new}{}}}
\subsection{Method \code{new()}}{
\code{mirai} launcher constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher$new(
  idle_time = NULL,
  wall_time = NULL,
  poll_high = NULL,
  poll_low = NULL,
  launch_timeout = NULL,
  launch_wait = NULL,
  max_tasks = NULL,
  async_dial = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{idle_time}}{Maximum number of seconds that a worker can idle
since the completion of the last task. If exceeded, the worker exits.}

\item{\code{wall_time}}{Soft wall time in seconds. See the \code{wall_time}
argument of \code{mirai::server()}.}

\item{\code{poll_high}}{High polling interval in seconds for the \code{mirai}
active queue.}

\item{\code{poll_low}}{Low polling interval in seconds for the \code{mirai}
active queue.}

\item{\code{launch_timeout}}{Number of seconds to time out
waiting for a new group of workers to launch.}

\item{\code{launch_wait}}{Number of seconds to wait between checks
that newly launched workers are ready to receive tasks.}

\item{\code{max_tasks}}{Maximum number of tasks that a worker will do before
exiting.}

\item{\code{async_dial}}{Logical, whether the \code{mirai} workers should dial in
asynchronously. See the \code{asyncdial} argument of \code{mirai::server()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{R6} object with the launcher.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the launcher.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
}
