% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_launcher.R
\name{crew_class_launcher}
\alias{crew_class_launcher}
\title{Launcher abstract class}
\description{
\code{R6} abstract class to build other subclasses
which launch and manage workers.
}
\seealso{
Other launchers: 
\code{\link{crew_class_launcher_callr}},
\code{\link{crew_launcher_callr}()}
}
\concept{launchers}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{workers}}{Named list of workers, where the names are the
sockets they connect to.}

\item{\code{seconds_start}}{Seconds of startup time to allow.
A worker is unconditionally assumed to be alive
from the moment of its launch until \code{seconds_start} seconds later.
After \code{seconds_start} seconds, the worker is only
considered alive if it is actively connected to its assign websocket.}

\item{\code{idle_time}}{Maximum number of seconds that a worker can idle
since the completion of the last task. If exceeded, the worker exits.}

\item{\code{wall_time}}{Soft wall time in seconds. See the \code{wall_time}
argument of \code{mirai::server()}.}

\item{\code{poll_high}}{High polling interval in seconds for the \code{mirai}
active queue.}

\item{\code{poll_low}}{Low polling interval in seconds for the \code{mirai}
active queue.}

\item{\code{launch_timeout}}{Number of seconds to time out
waiting for a new group of workers to launch.}

\item{\code{launch_wait}}{Number of seconds to wait between checks
that newly launched workers are ready to receive tasks.}

\item{\code{max_tasks}}{Maximum number of tasks that a worker will do before
exiting.}

\item{\code{async_dial}}{Logical, whether the \code{mirai} workers should dial in
asynchronously. See the \code{asyncdial} argument of \code{mirai::server()}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_launcher-new}{\code{crew_class_launcher$new()}}
\item \href{#method-crew_class_launcher-validate}{\code{crew_class_launcher$validate()}}
\item \href{#method-crew_class_launcher-starting}{\code{crew_class_launcher$starting()}}
\item \href{#method-crew_class_launcher-launch}{\code{crew_class_launcher$launch()}}
\item \href{#method-crew_class_launcher-terminate}{\code{crew_class_launcher$terminate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher-new}{}}}
\subsection{Method \code{new()}}{
\code{mirai} launcher constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher$new(
  seconds_start = NULL,
  idle_time = NULL,
  wall_time = NULL,
  poll_high = NULL,
  poll_low = NULL,
  launch_timeout = NULL,
  launch_wait = NULL,
  max_tasks = NULL,
  async_dial = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seconds_start}}{Seconds of startup time to allow.
A worker is unconditionally assumed to be alive
from the moment of its launch until \code{seconds_start} seconds later.
After \code{seconds_start} seconds, the worker is only
considered alive if it is actively connected to its assign websocket.}

\item{\code{idle_time}}{Maximum number of seconds that a worker can idle
since the completion of the last task. If exceeded, the worker exits.}

\item{\code{wall_time}}{Soft wall time in seconds. See the \code{wall_time}
argument of \code{mirai::server()}.}

\item{\code{poll_high}}{High polling interval in seconds for the \code{mirai}
active queue.}

\item{\code{poll_low}}{Low polling interval in seconds for the \code{mirai}
active queue.}

\item{\code{launch_timeout}}{Number of seconds to time out
waiting for a new group of workers to launch.}

\item{\code{launch_wait}}{Number of seconds to wait between checks
that newly launched workers are ready to receive tasks.}

\item{\code{max_tasks}}{Maximum number of tasks that a worker will do before
exiting.}

\item{\code{async_dial}}{Logical, whether the \code{mirai} workers should dial in
asynchronously. See the \code{asyncdial} argument of \code{mirai::server()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{R6} object with the launcher.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the launcher.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher-starting"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher-starting}{}}}
\subsection{Method \code{starting()}}{
Get the workers that recently launched
(within \code{seconds_start} seconds).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher$starting()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of worker websockets.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher-launch"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher-launch}{}}}
\subsection{Method \code{launch()}}{
Launch one or more workers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher$launch(sockets = character(0))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sockets}}{Sockets where the workers will dial in.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If a worker is already assigned to a socket,
the previous worker is terminated before the next
one is launched.
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_launcher-terminate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_launcher-terminate}{}}}
\subsection{Method \code{terminate()}}{
Terminate all workers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_launcher$terminate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
}
