% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_mirai_controller.R
\name{crew_class_mirai_controller}
\alias{crew_class_mirai_controller}
\title{\code{mirai} controller class}
\description{
\code{R6} class for \code{mirai} controllers.
}
\details{
See \code{\link[=crew_mirai_controller]{crew_mirai_controller()}}.
}
\concept{mirai-classes}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{router}}{Router object.}

\item{\code{launcher}}{Launcher object.}

\item{\code{tasks}}{Data frame of tasks in the queue.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_mirai_controller-new}{\code{crew_class_mirai_controller$new()}}
\item \href{#method-crew_class_mirai_controller-finalize}{\code{crew_class_mirai_controller$finalize()}}
\item \href{#method-crew_class_mirai_controller-validate}{\code{crew_class_mirai_controller$validate()}}
\item \href{#method-crew_class_mirai_controller-push}{\code{crew_class_mirai_controller$push()}}
\item \href{#method-crew_class_mirai_controller-pop}{\code{crew_class_mirai_controller$pop()}}
\item \href{#method-crew_class_mirai_controller-clone}{\code{crew_class_mirai_controller$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-new}{}}}
\subsection{Method \code{new()}}{
\code{mirai} controller constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$new(router = NULL, launcher = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{router}}{Router object. See \code{\link[=crew_mirai_controller]{crew_mirai_controller()}}.}

\item{\code{launcher}}{Launcher object. See \code{\link[=crew_mirai_controller]{crew_mirai_controller()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{R6} object with the controller object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-finalize}{}}}
\subsection{Method \code{finalize()}}{
Disconnect at garbage collection time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$finalize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the router.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-push"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-push}{}}}
\subsection{Method \code{push()}}{
Push a task to the head of the task list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$push(
  command,
  args = list(),
  timeout = NULL,
  name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{command}}{R code with the task to run.}

\item{\code{args}}{A list of objects referenced in \code{expr}.}

\item{\code{timeout}}{Time in milliseconds for the task.}

\item{\code{name}}{Optional name of the task. Replaced with a random name
if \code{NULL} or in conflict with an existing name in the task list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-pop"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-pop}{}}}
\subsection{Method \code{pop()}}{
Pop a completed task from as near as possible to the tail
of the task list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$pop()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
If there is a completed task available to collect, the return
value is a list object with the results, warnings, and errors.
Otherwise, if there are no tasks or if all tasks are still running,
the return value is \code{NULL}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_controller-clone"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_controller-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_controller$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
