% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_mirai_launcher_callr.R
\name{crew_class_mirai_launcher_callr}
\alias{crew_class_mirai_launcher_callr}
\title{\code{mirai} launcher class}
\description{
\code{R6} class to launch and manage \code{mirai} workers.
}
\details{
See \code{\link[=crew_mirai_launcher_callr]{crew_mirai_launcher_callr()}}.
}
\seealso{
Other launchers: 
\code{\link{crew_mirai_launcher_callr}()}
}
\concept{launchers}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sockets}}{TCP sockets for listening to the workers.}

\item{\code{workers}}{List of \code{callr::r_bg()} handles for workers.
The handle is \code{NA} if it is not yet called.}

\item{\code{idle}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{wall_time}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{poll_high}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{poll_low}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{poll_launch_timeout}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{poll_launch_wait}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{max_tasks}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{async_dial}}{Argument to \code{crew_mirai_launcher_callr()}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_mirai_launcher_callr-new}{\code{crew_class_mirai_launcher_callr$new()}}
\item \href{#method-crew_class_mirai_launcher_callr-validate}{\code{crew_class_mirai_launcher_callr$validate()}}
\item \href{#method-crew_class_mirai_launcher_callr-populate}{\code{crew_class_mirai_launcher_callr$populate()}}
\item \href{#method-crew_class_mirai_launcher_callr-running}{\code{crew_class_mirai_launcher_callr$running()}}
\item \href{#method-crew_class_mirai_launcher_callr-launch}{\code{crew_class_mirai_launcher_callr$launch()}}
\item \href{#method-crew_class_mirai_launcher_callr-terminate}{\code{crew_class_mirai_launcher_callr$terminate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_launcher_callr-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_launcher_callr-new}{}}}
\subsection{Method \code{new()}}{
\code{mirai} launcher constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_launcher_callr$new(
  sockets = NULL,
  workers = NULL,
  idle = NULL,
  wall_time = NULL,
  poll_high = NULL,
  poll_low = NULL,
  poll_launch_timeout = NULL,
  poll_launch_wait = NULL,
  max_tasks = NULL,
  async_dial = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sockets}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{workers}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{idle}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{wall_time}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{poll_high}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{poll_low}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{poll_launch_timeout}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{poll_launch_wait}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{max_tasks}}{Argument to \code{crew_mirai_launcher_callr()}.}

\item{\code{async_dial}}{Argument to \code{crew_mirai_launcher_callr()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{R6} object with the launcher.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_launcher_callr-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_launcher_callr-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the launcher.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_launcher_callr$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_launcher_callr-populate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_launcher_callr-populate}{}}}
\subsection{Method \code{populate()}}{
Populate workers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_launcher_callr$populate(sockets = character(0))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sockets}}{Character vector of local TCP sockets that the workers
will use to dial in to receive tasks.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_launcher_callr-running"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_launcher_callr-running}{}}}
\subsection{Method \code{running()}}{
Count the number of running workers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_launcher_callr$running()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Number of running workers.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_launcher_callr-launch"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_launcher_callr-launch}{}}}
\subsection{Method \code{launch()}}{
Launch one or more workers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_launcher_callr$launch(n = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Maximum number of workers to launch.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The actual number of newly launched workers
is less than or equal to the number of sockets
supplied to \verb{$populate()} minus the number of workers
already running.
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_launcher_callr-terminate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_launcher_callr-terminate}{}}}
\subsection{Method \code{terminate()}}{
Terminate all workers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_launcher_callr$terminate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
}
