% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_mirai_router.R
\name{crew_class_mirai_router}
\alias{crew_class_mirai_router}
\title{\code{mirai} router class}
\description{
\code{R6} class for \code{mirai} routers.
}
\details{
See \code{\link[=crew_mirai_router]{crew_mirai_router()}}.
}
\examples{

## ------------------------------------------------
## Method `crew_class_mirai_router$new`
## ------------------------------------------------

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_mirai_router()
router$validate()
router$connect()
# Now launch workers with a mirai launcher.
router$disconnect()
}
}
\seealso{
Other routers: 
\code{\link{crew_mirai_router}()}
}
\concept{routers}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{host}}{IP address the router uses to connect to the workers.}

\item{\code{port}}{TCP port the router uses to connect to the workers.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_mirai_router-new}{\code{crew_class_mirai_router$new()}}
\item \href{#method-crew_class_mirai_router-finalize}{\code{crew_class_mirai_router$finalize()}}
\item \href{#method-crew_class_mirai_router-validate}{\code{crew_class_mirai_router$validate()}}
\item \href{#method-crew_class_mirai_router-connected}{\code{crew_class_mirai_router$connected()}}
\item \href{#method-crew_class_mirai_router-connections}{\code{crew_class_mirai_router$connections()}}
\item \href{#method-crew_class_mirai_router-connect}{\code{crew_class_mirai_router$connect()}}
\item \href{#method-crew_class_mirai_router-disconnect}{\code{crew_class_mirai_router$disconnect()}}
\item \href{#method-crew_class_mirai_router-clone}{\code{crew_class_mirai_router$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-new}{}}}
\subsection{Method \code{new()}}{
\code{mirai} router constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$new(host = NULL, port = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{host}}{Host of the router. See \code{\link[=crew_mirai_router]{crew_mirai_router()}}.}

\item{\code{port}}{TCP port of the router. See \code{\link[=crew_mirai_router]{crew_mirai_router()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{R6} object with the router.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_mirai_router()
router$validate()
router$connect()
# Now launch workers with a mirai launcher.
router$disconnect()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-finalize}{}}}
\subsection{Method \code{finalize()}}{
Disconnect at garbage collection time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$finalize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the router.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-connected"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-connected}{}}}
\subsection{Method \code{connected()}}{
Check if the router is connected.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$connected()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if connected and \code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-connections"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-connections}{}}}
\subsection{Method \code{connections()}}{
Return the number of workers connected to the router.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$connections()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nonnegative integer, number of connected workers.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-connect"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-connect}{}}}
\subsection{Method \code{connect()}}{
Connect the router to the host and port.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$connect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnect the router from the host and port.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$disconnect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-clone"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
