% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_mirai_router.R
\name{crew_class_mirai_router}
\alias{crew_class_mirai_router}
\title{\code{mirai} router class}
\description{
\code{R6} class for \code{mirai} routers.
}
\details{
See \code{\link[=crew_mirai_router]{crew_mirai_router()}}.
}
\examples{

## ------------------------------------------------
## Method `crew_class_mirai_router$new`
## ------------------------------------------------

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_mirai_router()
router$validate()
router$connect()
# Now launch workers with a mirai launcher.
router$disconnect()
}
}
\seealso{
Other routers: 
\code{\link{crew_mirai_router}()}
}
\concept{routers}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of the router.}

\item{\code{host}}{Client IP address. See \code{\link[=crew_mirai_router]{crew_mirai_router()}} for details.}

\item{\code{ports}}{TCP ports. See \code{\link[=crew_mirai_router]{crew_mirai_router()}} for details.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_mirai_router-new}{\code{crew_class_mirai_router$new()}}
\item \href{#method-crew_class_mirai_router-finalize}{\code{crew_class_mirai_router$finalize()}}
\item \href{#method-crew_class_mirai_router-validate}{\code{crew_class_mirai_router$validate()}}
\item \href{#method-crew_class_mirai_router-sockets}{\code{crew_class_mirai_router$sockets()}}
\item \href{#method-crew_class_mirai_router-available_sockets}{\code{crew_class_mirai_router$available_sockets()}}
\item \href{#method-crew_class_mirai_router-connected}{\code{crew_class_mirai_router$connected()}}
\item \href{#method-crew_class_mirai_router-connections}{\code{crew_class_mirai_router$connections()}}
\item \href{#method-crew_class_mirai_router-connect}{\code{crew_class_mirai_router$connect()}}
\item \href{#method-crew_class_mirai_router-disconnect}{\code{crew_class_mirai_router$disconnect()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-new}{}}}
\subsection{Method \code{new()}}{
\code{mirai} router constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$new(name = NULL, host = NULL, ports = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the router object.}

\item{\code{host}}{Client IP address. See \code{\link[=crew_mirai_router]{crew_mirai_router()}} for details.}

\item{\code{ports}}{TCP ports. See \code{\link[=crew_mirai_router]{crew_mirai_router()}} for details.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{R6} object with the router.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_mirai_router()
router$validate()
router$connect()
# Now launch workers with a mirai launcher.
router$disconnect()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-finalize}{}}}
\subsection{Method \code{finalize()}}{
Disconnect at garbage collection time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$finalize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the router.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-sockets"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-sockets}{}}}
\subsection{Method \code{sockets()}}{
Get the TCP socket of the router.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$sockets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string with the TCP socket of the router.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-available_sockets"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-available_sockets}{}}}
\subsection{Method \code{available_sockets()}}{
Get available TCP sockets for listening to workers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$available_sockets(n = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Maximum number of sockets to return.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character string with available TCP sockets.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-connected"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-connected}{}}}
\subsection{Method \code{connected()}}{
Check if the router is connected.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$connected()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if successfully listening for dialed-in workers,
\code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-connections"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-connections}{}}}
\subsection{Method \code{connections()}}{
Show worker connections.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$connections()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of TCP sockets where the client is
listening to currently dialed-in running workers.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-connect"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-connect}{}}}
\subsection{Method \code{connect()}}{
Start listening for workers on the available sockets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$connect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnect the router.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$disconnect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
}
