% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_mirai_router.R
\name{crew_class_mirai_router}
\alias{crew_class_mirai_router}
\title{\code{mirai} router class}
\description{
\code{R6} class for \code{mirai} routers.
}
\details{
See \code{\link[=crew_mirai_router]{crew_mirai_router()}}.
}
\seealso{
Other routers: 
\code{\link{crew_mirai_router}()}
}
\concept{routers}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of the router.}

\item{\code{host}}{Local IP address.}

\item{\code{ports}}{TCP ports to listen to workers.}

\item{\code{timeout}}{Timeout in seconds (dis)connecting the \code{mirai} router.}

\item{\code{wait}}{Polling interval in seconds checking if the \code{mirai}
router successfully (dis)connected.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_mirai_router-new}{\code{crew_class_mirai_router$new()}}
\item \href{#method-crew_class_mirai_router-validate}{\code{crew_class_mirai_router$validate()}}
\item \href{#method-crew_class_mirai_router-sockets_listening}{\code{crew_class_mirai_router$sockets_listening()}}
\item \href{#method-crew_class_mirai_router-sockets_occupied}{\code{crew_class_mirai_router$sockets_occupied()}}
\item \href{#method-crew_class_mirai_router-sockets_available}{\code{crew_class_mirai_router$sockets_available()}}
\item \href{#method-crew_class_mirai_router-is_connected}{\code{crew_class_mirai_router$is_connected()}}
\item \href{#method-crew_class_mirai_router-connect}{\code{crew_class_mirai_router$connect()}}
\item \href{#method-crew_class_mirai_router-disconnect}{\code{crew_class_mirai_router$disconnect()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-new}{}}}
\subsection{Method \code{new()}}{
\code{mirai} router constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$new(
  name = NULL,
  host = NULL,
  ports = NULL,
  timeout = NULL,
  wait = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Argument passed from \code{\link[=crew_mirai_router]{crew_mirai_router()}}.}

\item{\code{host}}{Argument passed from \code{\link[=crew_mirai_router]{crew_mirai_router()}}.}

\item{\code{ports}}{Argument passed from \code{\link[=crew_mirai_router]{crew_mirai_router()}}.}

\item{\code{timeout}}{Argument passed from \code{\link[=crew_mirai_router]{crew_mirai_router()}}.}

\item{\code{wait}}{Argument passed from \code{\link[=crew_mirai_router]{crew_mirai_router()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{R6} object with the router.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the router.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-sockets_listening"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-sockets_listening}{}}}
\subsection{Method \code{sockets_listening()}}{
Get all sockets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$sockets_listening()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of all sockets if connected, or
\code{character(0)} if the router is not connected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-sockets_occupied"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-sockets_occupied}{}}}
\subsection{Method \code{sockets_occupied()}}{
Show worker connections.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$sockets_occupied()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of TCP sockets where workers
are currently dialed in.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-sockets_available"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-sockets_available}{}}}
\subsection{Method \code{sockets_available()}}{
Get TCP sockets that are available for workers to dial in.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$sockets_available()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string with available TCP sockets.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-is_connected"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-is_connected}{}}}
\subsection{Method \code{is_connected()}}{
Check if the router is connected.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$is_connected()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if successfully listening for dialed-in workers,
\code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-connect"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-connect}{}}}
\subsection{Method \code{connect()}}{
Start listening for workers on the available sockets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$connect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_mirai_router-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_mirai_router-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnect the router.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_mirai_router$disconnect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
}
