% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_router.R
\name{crew_class_router}
\alias{crew_class_router}
\title{Router class}
\description{
\code{R6} class for \code{mirai} routers.
}
\details{
See \code{\link[=crew_router]{crew_router()}}.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_router()
router$start()
router$daemons
router$terminate()
}

## ------------------------------------------------
## Method `crew_class_router$new`
## ------------------------------------------------

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_router()
router$start()
router$daemons
router$terminate()
}
}
\seealso{
Other routers: 
\code{\link{crew_router}()}
}
\concept{routers}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{See \code{\link[=crew_router]{crew_router()}}.}

\item{\code{workers}}{See \code{\link[=crew_router]{crew_router()}}.}

\item{\code{host}}{See \code{\link[=crew_router]{crew_router()}}.}

\item{\code{port}}{See \code{\link[=crew_router]{crew_router()}}.}

\item{\code{seconds_interval}}{See \code{\link[=crew_router]{crew_router()}}.}

\item{\code{seconds_timeout}}{See \code{\link[=crew_router]{crew_router()}}.}

\item{\code{started}}{Whether the router is started.}

\item{\code{dispatcher}}{Process ID of the \code{mirai} dispatcher}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_router-new}{\code{crew_class_router$new()}}
\item \href{#method-crew_class_router-validate}{\code{crew_class_router$validate()}}
\item \href{#method-crew_class_router-start}{\code{crew_class_router$start()}}
\item \href{#method-crew_class_router-log}{\code{crew_class_router$log()}}
\item \href{#method-crew_class_router-terminate}{\code{crew_class_router$terminate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_router-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_router-new}{}}}
\subsection{Method \code{new()}}{
\code{mirai} router constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_router$new(
  name = NULL,
  workers = NULL,
  host = NULL,
  port = NULL,
  seconds_interval = NULL,
  seconds_timeout = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Argument passed from \code{\link[=crew_router]{crew_router()}}.}

\item{\code{workers}}{Argument passed from \code{\link[=crew_router]{crew_router()}}.}

\item{\code{host}}{Argument passed from \code{\link[=crew_router]{crew_router()}}.}

\item{\code{port}}{Argument passed from \code{\link[=crew_router]{crew_router()}}.}

\item{\code{seconds_interval}}{Argument passed from \code{\link[=crew_router]{crew_router()}}.}

\item{\code{seconds_timeout}}{Argument passed from \code{\link[=crew_router]{crew_router()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{R6} object with the router.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_router()
router$start()
router$daemons
router$terminate()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_router-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_router-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the router.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_router$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_router-start"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_router-start}{}}}
\subsection{Method \code{start()}}{
Start listening for workers on the available sockets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_router$start()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_router-log"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_router-log}{}}}
\subsection{Method \code{log()}}{
Show an informative worker log.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_router$log()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{tibble} with information on the workers.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_router-terminate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_router-terminate}{}}}
\subsection{Method \code{terminate()}}{
Stop the mirai client and disconnect from the
worker websockets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_router$terminate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
}
