% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_router.R
\name{crew_class_router}
\alias{crew_class_router}
\title{Router class}
\description{
\code{R6} class for \code{mirai} routers.
}
\details{
See \code{\link[=crew_router]{crew_router()}}.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_router()
router$sockets_listening() # character(0)
router$connect()
router$sockets_listening() # "ws://xx.xx.xx:xxxxx/x"
router$disconnect()
router$sockets_listening() # character(0)
}

## ------------------------------------------------
## Method `crew_class_router$new`
## ------------------------------------------------

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_router()
router$sockets_listening() # character(0)
router$connect()
router$sockets_listening() # "ws://xx.xx.xx:xxxxx/x"
router$disconnect()
router$sockets_listening() # character(0)
}
}
\seealso{
Other routers: 
\code{\link{crew_router}()}
}
\concept{routers}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of the router.}

\item{\code{sockets}}{Websockets to listen for workers.}

\item{\code{router_timeout}}{Timeout in seconds for checking the \code{mirai}
client connection.}

\item{\code{router_wait}}{Polling interval in seconds for checking the
\code{mirai} client connection.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_router-new}{\code{crew_class_router$new()}}
\item \href{#method-crew_class_router-validate}{\code{crew_class_router$validate()}}
\item \href{#method-crew_class_router-occupied}{\code{crew_class_router$occupied()}}
\item \href{#method-crew_class_router-unoccupied}{\code{crew_class_router$unoccupied()}}
\item \href{#method-crew_class_router-connected}{\code{crew_class_router$connected()}}
\item \href{#method-crew_class_router-connect}{\code{crew_class_router$connect()}}
\item \href{#method-crew_class_router-disconnect}{\code{crew_class_router$disconnect()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_router-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_router-new}{}}}
\subsection{Method \code{new()}}{
\code{mirai} router constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_router$new(
  name = NULL,
  sockets = NULL,
  router_timeout = NULL,
  router_wait = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Argument passed from \code{\link[=crew_router]{crew_router()}}.}

\item{\code{sockets}}{Argument passed from \code{\link[=crew_router]{crew_router()}}.}

\item{\code{router_timeout}}{Argument passed from \code{\link[=crew_router]{crew_router()}}.}

\item{\code{router_wait}}{Argument passed from \code{\link[=crew_router]{crew_router()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{R6} object with the router.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_router()
router$sockets_listening() # character(0)
router$connect()
router$sockets_listening() # "ws://xx.xx.xx:xxxxx/x"
router$disconnect()
router$sockets_listening() # character(0)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_router-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_router-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the router.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_router$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_router-occupied"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_router-occupied}{}}}
\subsection{Method \code{occupied()}}{
Get occupied worker sockets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_router$occupied()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of websockets sockets where workers
are currently dialed in.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_router-unoccupied"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_router-unoccupied}{}}}
\subsection{Method \code{unoccupied()}}{
Get worker sockets that are unoccupied and
available for workers to dial in.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_router$unoccupied()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string with available websockets.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_router-connected"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_router-connected}{}}}
\subsection{Method \code{connected()}}{
Check if the router is connected.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_router$connected()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method may stall and time out if there are
tasks in the queue. Methods \code{connect()} and \code{disconnect()}
call \code{connected()} to manage the connection before
and after the entire workload, respectively.
}

\subsection{Returns}{
\code{TRUE} if successfully listening for dialed-in workers,
\code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_router-connect"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_router-connect}{}}}
\subsection{Method \code{connect()}}{
Start listening for workers on the available sockets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_router$connect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_router-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_router-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnect the router.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_router$disconnect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
}
