% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_clean.R
\name{crew_clean}
\alias{crew_clean}
\title{Deprecated: terminate dispatchers and/or workers}
\usage{
crew_clean(
  dispatchers = TRUE,
  workers = TRUE,
  user = ps::ps_username(),
  seconds_interval = 0.25,
  seconds_timeout = 60,
  verbose = TRUE
)
}
\arguments{
\item{dispatchers}{Logical of length 1,
whether to terminate dispatcher processes
(in older versions of \code{mirai}: <= 2.6.1).
In \code{mirai} > 2.6.1, the dispatcher is a thread within
the controller's own R process
and will not be terminated by this function.}

\item{workers}{Logical of length 1,
whether to terminate workers.}

\item{user}{Character of length 1. Terminate processes
associated with this user name.}

\item{seconds_interval}{Seconds to wait between polling intervals
waiting for a process to exit.}

\item{seconds_timeout}{Seconds to wait for a process to exit.}

\item{verbose}{Logical of length 1, whether to print an informative
message every time a process is terminated.}
}
\value{
\code{NULL} (invisibly). If \code{verbose} is \code{TRUE}, it does
print out a message for every terminated process.
}
\description{
Deprecated on 2025-08-26 in \code{crew} version 1.2.1.9006.
Please use \code{\link[=crew_monitor_local]{crew_monitor_local()}} instead.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_clean()
}
}
\seealso{
Other utility: 
\code{\link{crew_assert}()},
\code{\link{crew_deprecate}()},
\code{\link{crew_eval}()},
\code{\link{crew_random_name}()},
\code{\link{crew_retry}()},
\code{\link{crew_terminate_process}()},
\code{\link{crew_terminate_signal}()},
\code{\link{crew_worker}()}
}
\concept{utility}
