% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_clean.R
\name{crew_clean}
\alias{crew_clean}
\title{Deprecated: terminate dispatchers and/or workers}
\usage{
crew_clean(
  dispatchers = TRUE,
  workers = TRUE,
  user = ps::ps_username(),
  seconds_interval = 0.25,
  seconds_timeout = 60,
  verbose = TRUE
)
}
\arguments{
\item{dispatchers}{Logical of length 1,
whether to terminate dispatchers.}

\item{workers}{Logical of length 1,
whether to terminate workers.}

\item{user}{Character of length 1. Terminate dispatchers and/or
workers associated with this user name.}

\item{seconds_interval}{Seconds to wait between polling intervals
waiting for a process to exit.}

\item{seconds_timeout}{Seconds to wait for a process to exit.}

\item{verbose}{Logical of length 1, whether to print an informative
message every time a process is terminated.}
}
\value{
\code{NULL} (invisibly). If \code{verbose} is \code{TRUE}, it does
print out a message for every terminated process.
}
\description{
Deprecated on 2025-08-26 in \code{crew} version 1.2.1.9006.
Please use \code{\link[=crew_monitor_local]{crew_monitor_local()}} instead.
}
\details{
Behind the scenes, \code{mirai} uses an external R process
called a "dispatcher" to send tasks to \code{crew} workers.
This dispatcher usually shuts down when you terminate the controller
or quit your R session, but sometimes it lingers. Likewise,
sometimes \code{crew} workers do not shut down on their own.
The \code{crew_clean()} function searches the process table on your
local machine and manually terminates any \code{mirai} dispatchers
and \code{crew} workers associated with your user name (or the
user name you select in the \code{user} argument.
Unfortunately, it cannot reach remote workers such as those
launched by a \code{crew.cluster} controller.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_clean()
}
}
\seealso{
Other utility: 
\code{\link{crew_assert}()},
\code{\link{crew_deprecate}()},
\code{\link{crew_eval}()},
\code{\link{crew_random_name}()},
\code{\link{crew_retry}()},
\code{\link{crew_terminate_process}()},
\code{\link{crew_terminate_signal}()},
\code{\link{crew_worker}()}
}
\concept{utility}
