% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_controller_sequential.R
\name{crew_controller_sequential}
\alias{crew_controller_sequential}
\title{Create a sequential controller.}
\usage{
crew_controller_sequential()
}
\description{
The sequential controller runs tasks on the same R process
where the controller object exists. Tasks run sequentially
rather than in parallel.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
controller <- crew_controller_sequential()
controller$push(name = "task", command = sqrt(4))
controller$pop()
}
}
\seealso{
Other sequential controllers: 
\code{\link{crew_class_controller_sequential}}
}
\concept{sequential controllers}
