% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_worker.R
\name{crew_log}
\alias{crew_log}
\title{Print a \code{\link[=crew_worker]{crew_worker()}} log message.}
\usage{
crew_log(..., log = NULL)
}
\arguments{
\item{...}{Character strings to print.}

\item{log}{Character of length 1 with the path
to a log file. If \code{NULL}, then no log is written.}
}
\value{
Nothing (invisibly) but print a message.
}
\description{
Not a user-side function. Do not invoke directly.
}
\examples{
log <- tempfile()
crew_log("done", log = log)
writeLines(readLines(log))
}
\keyword{internal}
