% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_controller_mirai.R
\name{crew_mirai_controller}
\alias{crew_mirai_controller}
\title{Create a \code{mirai} controller.}
\usage{
crew_mirai_controller(router, launcher)
}
\arguments{
\item{router}{An \code{R6} router object created by \code{\link[=crew_mirai_router]{crew_mirai_router()}}.}

\item{launcher}{An \code{R6} launcher object created by one of the
\verb{crew_mirai_launcher_*()} functions.}
}
\description{
Create an \code{R6} object to submit tasks and launch workers
on \code{mirai}-based infrastructure.
}
\examples{
stop("TBD")
}
\concept{mirai}
