% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_mirai_launcher_callr.R
\name{crew_mirai_launcher_callr}
\alias{crew_mirai_launcher_callr}
\title{Create a \code{mirai} launcher with \code{callr} workers.}
\usage{
crew_mirai_launcher_callr(
  sockets,
  seconds_idle = Inf,
  seconds_wall_time = Inf,
  seconds_poll_high = 5L,
  seconds_poll_low = 50L,
  max_tasks = Inf,
  async_dial = TRUE
)
}
\arguments{
\item{sockets}{TCP sockets for listening to the workers.}

\item{seconds_idle}{Maximum number of seconds that a worker can idle
since the completion of the last task. If exceeded, the worker exits.}

\item{seconds_wall_time}{Soft wall time in seconds. See the \code{wall_time}
argument of \code{mirai::server()}.}

\item{seconds_poll_high}{High polling interval in seconds for the \code{mirai}
active queue.}

\item{seconds_poll_low}{Low polling interval in seconds for the \code{mirai}
active queue.}

\item{async_dial}{Logical, whether the \code{mirai} workers should dial in
asynchronously. See the \code{asyncdial} argument of \code{mirai::server()}.}
}
\description{
Create an \code{R6} object to launch and maintain
\code{callr} workers for a \code{mirai} controller.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
launcher <- crew_mirai_launcher_callr()
launcher$validate()
}
}
\seealso{
Other launchers: 
\code{\link{crew_class_mirai_launcher_callr}}
}
\concept{launchers}
