% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_mirai_router.R
\name{crew_mirai_router}
\alias{crew_mirai_router}
\title{Create a \code{mirai} router.}
\usage{
crew_mirai_router(name = NULL, host = NULL, ports = NULL)
}
\arguments{
\item{name}{Name of the \code{mirai} router.
Defaults to a string from \code{ids::proquint()}.}

\item{host}{IP address of the client process that the workers can dial
into inside the local network.
If a character string, the router uses the specified IP address.
If \code{NULL}, the IP address defaults to \code{getip::getip(type = "local")}.}

\item{ports}{Integer vector of TCP port numbers, one for each port
on the client where a worker (\code{mirai} server) may dial in and accept tasks.
For random available ports, you can use \code{parallelly::freePort()}.
Defaults to a single port from \code{parallelly::freePort()}
in the ephemeral range (49152 to 65535).}
}
\description{
Create an \code{R6} object to route tasks to \code{mirai} workers
on the local network.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_mirai_router()
router$validate()
router$connect()
# Now launch workers with a mirai launcher.
router$disconnect()
}
}
\seealso{
Other routers: 
\code{\link{crew_class_mirai_router}}
}
\concept{routers}
