% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_multi_controller.R
\name{crew_multi_controller}
\alias{crew_multi_controller}
\title{Create a multi-controller.}
\usage{
crew_multi_controller(...)
}
\arguments{
\item{...}{\code{R6} controller objects or lists of \code{R6} controller objects.
Nested lists are allowed, but each element must be a control object
or another list.}
}
\description{
Create an \code{R6} object to submit tasks and launch workers
through multiple \code{crew} controllers.
}
\seealso{
Other controllers: 
\code{\link{crew_class_multi_controller}},
\code{\link{crew_mirai_controller_callr}()}
}
\concept{controllers}
\keyword{internal}
