% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_queue.R
\name{crew_queue}
\alias{crew_queue}
\title{Create a \code{crew} queue object.}
\usage{
crew_queue(step = 1000L)
}
\arguments{
\item{step}{Positive integer with the number of elements to extend the
queue on each call to the \code{extend()} method.}
}
\value{
A queue object.
}
\description{
Create an \code{R6} \code{crew} queue object.
}
\details{
A \code{crew} queue object efficiently tracks the names of resolved
tasks so the controller can pop them efficiently.
It is also used to track backlogged tasks.
}
\seealso{
Other queue: 
\code{\link{crew_class_queue}}
}
\concept{queue}
