% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_queue.R
\name{crew_queue}
\alias{crew_queue}
\title{Create a \code{crew} queue object.}
\usage{
crew_queue(data = character(0L), step = 1000L)
}
\arguments{
\item{data}{Character vector of initial queue data.}

\item{step}{Positive integer with the number of elements to extend the
queue on each call to the \code{extend()} method.}
}
\value{
A queue object.
}
\description{
Create an \code{R6} \code{crew} queue object.
}
\details{
A \code{crew} queue is a classical first-in-first-out data structure
that extends itself in chunks (of size \code{step}) to avoid
overhead.
\code{crew} uses queues to efficiently track the names of resolved
tasks and backlogged tasks.
}
\seealso{
Other queue: 
\code{\link{crew_class_queue}}
}
\concept{queue}
