% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_queue.R
\name{crew_queue}
\alias{crew_queue}
\title{Create a \code{crew} queue object.}
\usage{
crew_queue()
}
\description{
Create an \code{R6} \code{crew} queue object for resolved task names.
}
\details{
A \code{crew} queue object efficiently tracks the names of resolved
tasks so the controller can pop them efficiently.
}
\seealso{
Other queue: 
\code{\link{crew_class_queue}}
}
\concept{queue}
