% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_queue_bg.R
\name{crew_queue_bg}
\alias{crew_queue_bg}
\title{Local background queue}
\description{
Task queue with \code{callr::r_bg()} workers.
}
\details{
Some public methods are inherited from \link{crew_queue}.
See the \link{crew_queue} help file for details on those.

The interface is the same as \link{crew_queue_session} except
there is a data store (e.g. \link{crew_store_local}) which
the user can customize. Most users should choose for the session queue
\link{crew_queue_session} instead of the background queue
\link{crew_queue_bg}. The only reason to use \link{crew_queue_bg}
is for subqueues for \link{crew_queue_future} on clusters
where \code{batchtools} futures clash with \code{callr::r_session}.
}
\section{Queue attribution}{

The design and implementation of \code{crew} task queues, especially
the local \code{callr}-based task queues, borrows heavily from
the work of Gábor Csárdi, particularly
\url{https://www.tidyverse.org/blog/2019/09/callr-task-q/},
which was also released in commit
\verb{811a02f604de2cf03264f6b35ce9ec8a412f2581}
of \url{https://github.com/r-lib/callr} in the
\code{vignettes/Task-queue-with-callr.Rmd}
file under the MIT license. See the \code{crew} package \code{README.md} and
\code{NOTICE} files for other mentions of this attribution.
}

\examples{
fun <- function(x) x + 1
args <- list(x = 1)
queue <- crew_queue_bg$new(timeout = 60, wait = 0.1)
queue$push(fun = fun, args = args)
queue$block()
result <- queue$pop()
str(result)
result$result$result
queue$shutdown()
processx::supervisor_kill()
}
\seealso{
Other queue: 
\code{\link{crew_queue_future}},
\code{\link{crew_queue_session}},
\code{\link{crew_queue}}
}
\concept{queue}
\section{Super class}{
\code{\link[crew:crew_queue]{crew::crew_queue}} -> \code{crew_queue_bg}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_queue_bg-new}{\code{crew_queue_bg$new()}}
\item \href{#method-crew_queue_bg-shutdown}{\code{crew_queue_bg$shutdown()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_queue" data-id="block"><a href='../../crew/html/crew_queue.html#method-crew_queue-block'><code>crew::crew_queue$block()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_queue" data-id="crashed"><a href='../../crew/html/crew_queue.html#method-crew_queue-crashed'><code>crew::crew_queue$crashed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_queue" data-id="get_results"><a href='../../crew/html/crew_queue.html#method-crew_queue-get_results'><code>crew::crew_queue$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_queue" data-id="get_tasks"><a href='../../crew/html/crew_queue.html#method-crew_queue-get_tasks'><code>crew::crew_queue$get_tasks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_queue" data-id="get_workers"><a href='../../crew/html/crew_queue.html#method-crew_queue-get_workers'><code>crew::crew_queue$get_workers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_queue" data-id="pop"><a href='../../crew/html/crew_queue.html#method-crew_queue-pop'><code>crew::crew_queue$pop()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_queue" data-id="push"><a href='../../crew/html/crew_queue.html#method-crew_queue-push'><code>crew::crew_queue$push()</code></a></span></li>
<li><span class="pkg-link" data-pkg="crew" data-topic="crew_queue" data-id="update"><a href='../../crew/html/crew_queue.html#method-crew_queue-update'><code>crew::crew_queue$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_queue_bg-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_queue_bg-new}{}}}
\subsection{Method \code{new()}}{
Local background queue constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_queue_bg$new(
  workers = 1,
  timeout = 60,
  wait = 0.1,
  store = crew_store_local$new(timeout = Inf, wait = wait)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workers}}{Number of workers in the queue.}

\item{\code{timeout}}{Number of seconds to for a worker to wait
for something to happen (e.g. the arrival of a task)
before timing out and quitting.}

\item{\code{wait}}{Number of seconds to wait in between iterations while
waiting for something to happen (e.g. the arrival of a task).}

\item{\code{store}}{A local store object from \link{crew_store_local}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A local background queue object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_queue_bg-shutdown"></a>}}
\if{latex}{\out{\hypertarget{method-crew_queue_bg-shutdown}{}}}
\subsection{Method \code{shutdown()}}{
Shut down the workers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_queue_bg$shutdown()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly)
}
}
}
