% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_random_name.R
\name{crew_random_name}
\alias{crew_random_name}
\title{Random name}
\usage{
crew_random_name(n = 12L)
}
\arguments{
\item{n}{Number of bytes of information in the random string
hashed to generate the name. Larger \code{n} is
more likely to generate unique names, but it may be slower to compute.}
}
\value{
A random character string.
}
\description{
Generate a random string that can be used as a name
for a worker or task.
}
\details{
The randomness is not reproducible and cannot be set with
e.g. \code{set.seed()} in R.
}
\examples{
crew_random_name()
}
\seealso{
Other utilities: 
\code{\link{crew_assert}()},
\code{\link{crew_deprecate}()},
\code{\link{crew_eval_async}()},
\code{\link{crew_eval}()},
\code{\link{crew_retry}()},
\code{\link{crew_worker}()}
}
\concept{utilities}
