% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_router.R
\name{crew_router}
\alias{crew_router}
\title{Create a router.}
\usage{
crew_router(
  name = NULL,
  workers = 1L,
  port = 0L,
  router_timeout = 5,
  router_wait = 0.1
)
}
\arguments{
\item{name}{Name of the router object. If \code{NULL}, a name is automatically
generated.}

\item{workers}{Integer, maximum number of parallel workers to run.}

\item{port}{TCP port to listen for the workers. If \code{0},
then NNG automatically chooses an available ephemeral port.}

\item{router_timeout}{Number of seconds to time out waiting for the \code{mirai}
client to (dis)connect.}

\item{router_wait}{Number of seconds to wait between iterations checking
if the \code{mirai} client is (dis)connected.}
}
\description{
Create an \code{R6} object to manage the \code{mirai} task scheduler
client.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_router()
router$listen()
router$sockets() # "ws://xx.xx.xx:xxxxx/x"
router$terminate()
}
}
\seealso{
Other routers: 
\code{\link{crew_class_router}}
}
\concept{routers}
\keyword{internal}
