% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_router.R
\name{crew_router}
\alias{crew_router}
\title{Create a router.}
\usage{
crew_router(
  name = NULL,
  workers = 1L,
  host = NULL,
  ports = NULL,
  router_timeout = 5,
  router_wait = 0.1
)
}
\arguments{
\item{name}{Name of the \code{mirai} router.
Defaults to a string from \code{ids::proquint()}.}

\item{workers}{Integer, maximum number of parallel workers to run.
\code{crew} will reserve one ephemeral port for each worker. See the
Ports section for an important cautionary note.}

\item{host}{IP address of the client process that the workers can dial
into inside the local network.
If a character string, the router uses the specified IP address.
If \code{NULL}, the IP address defaults to \code{getip::getip(type = "local")}.}

\item{ports}{Optional integer vector of TCP port numbers.
Supersedes \code{workers} if supplied. Except for 0,
which defers to NNG to automatically assign a port, each unique port
value corresponds to a local port where a worker will dial in
to accept tasks.}

\item{router_timeout}{Number of seconds to time out waiting for the \code{mirai}
router to (dis)connect.}

\item{router_wait}{Number of seconds to wait between iterations checking
if the \code{mirai} router is (dis)connected.}
}
\description{
Create an \code{R6} object to route tasks to workers
on the local network.
}
\section{Ports}{

In the \code{mirai}-based task scheduling in \code{crew}, each parallel worker
dials into a different TCP port on the local machine.
If you launch hundreds of workers, then hundreds of ports will
not be available to other users or processes.
Large numbers of workers on shared machines or clusters may
seriously disrupt the tasks of other users, so please be careful.
}

\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_router()
router$sockets_listening() # character(0)
router$connect()
router$sockets_listening() # "tcp://xx.xx.xx:xxxxx"
router$disconnect()
router$sockets_listening() # character(0)
}
}
\seealso{
Other routers: 
\code{\link{crew_class_router}}
}
\concept{routers}
\keyword{internal}
