% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_router.R
\name{crew_router}
\alias{crew_router}
\title{Create a router.}
\usage{
crew_router(
  name = NULL,
  workers = 1L,
  host = NULL,
  port = NULL,
  seconds_interval = 0.01,
  seconds_timeout = 5
)
}
\arguments{
\item{name}{Name of the router object. If \code{NULL}, a name is automatically
generated.}

\item{workers}{Integer, maximum number of parallel workers to run.}

\item{host}{IP address of the \code{mirai} client to send and receive tasks.
If \code{NULL}, the host defaults to the local IP address.}

\item{port}{TCP port to listen for the workers. If \code{NULL},
then an available ephemeral port is automatically chosen.}

\item{seconds_interval}{Number of seconds between
polling intervals waiting for certain internal
synchronous operations to complete.}

\item{seconds_timeout}{Number of seconds until timing
out while waiting for certain synchronous operations to complete.}
}
\description{
Create an \code{R6} object to manage the \code{mirai} task scheduler
client.
}
\examples{
if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_router()
router$start()
router$daemons
router$terminate()
}
}
\seealso{
Other routers: 
\code{\link{crew_class_router}}
}
\concept{routers}
\keyword{internal}
