% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_store_local.R
\name{crew_store_local}
\alias{crew_store_local}
\title{Local store}
\description{
Local store class.
}
\details{
Most \code{crew} queues use data stores to send and receive task data.
Some use entirely local data, others may use cloud data to support
cloud workers. The local store works with files on the local file system.
}
\examples{
# Usage on its own:
store <- crew_store_local$new(timeout = Inf)
store$write_input("worker", list(value = "x"))
store$list_input()
store$exists_input("worker")
store$read_input("worker")
# Usage in a queue:
fun <- function(x) x + 1
args <- list(x = 1)
queue <- crew_queue_future$new(timeout = 60, store = store)
queue$push(fun = fun, args = args)
queue$block()
result <- queue$pop()
str(result)
result$result$result
queue$shutdown()
processx::supervisor_kill()
store$destroy()
}
\seealso{
Other store: 
\code{\link{crew_store}}
}
\concept{store}
\section{Super class}{
\code{\link[crew:crew_store]{crew::crew_store}} -> \code{crew_store_local}
}
