% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_worker.R
\name{crew_task}
\alias{crew_task}
\title{Run a task in a crew worker.}
\usage{
crew_task(worker, store, timeout, wait, log = NULL)
}
\arguments{
\item{worker}{Character of length 1, name of the worker.}

\item{store}{Either a data store object or a character string
with a marshaled data store.}

\item{timeout}{Positive numeric of length 1, number of seconds
that a worker can idle before timing out.}

\item{wait}{Positive numeric of length 1, number of seconds
that the worker waits between checking if a task exists.}
}
\description{
Not a user-side function. Do not invoke directly.
}
\examples{
if (!identical(Sys.getenv("CREW_EXAMPLES", unset = ""), "")) {
root <- tempfile()
dir.create(root)
store <- crew_store_local$new(root = root)
fun <- function(x) {
  x + 1
}
args <- list(x = 1)
value <- list(fun = deparse(fun), args = args)
store$write_input(worker = "my_worker", value = value)
log <- tempfile()
crew_task(
  worker = "my_worker",
  store = store$marshal(),
  timeout = 0,
  wait = 0,
  log = log
)
store$read_output("my_worker")$result # 2
writeLines(readLines(log))
}
}
\keyword{internal}
