% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_worker.R
\name{crew_worker}
\alias{crew_worker}
\title{Worker event loop.}
\usage{
crew_worker(worker, store, jobs, timeout, wait)
}
\arguments{
\item{jobs}{Maximum number of jobs before returning.}

\item{timeout}{Positive numeric of length 1, number of seconds
that a worker can idle before timing out.}

\item{wait}{Positive numeric of length 1, number of seconds
that the worker waits between checking if a job exists.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Not a user-side function. Do not invoke directly.
}
\details{
The worker event loop runs inside the underlying process
of a worker. It waits for the next job (timing out
if it idles for too long at one time) runs any incoming jobs
and puts the output in the data store.
}
\examples{
if (!identical(Sys.getenv("CREW_EXAMPLES", unset = ""), "")) {
dir_root <- tempfile()
dir.create(dir_root)
store <- store_local$new(dir_root = dir_root)
fun <- function(x) {
  x + 1
}
args <- list(x = 1)
value <- list(fun = deparse(fun), args = args)
store$write_worker_input(worker = "my_worker", value = value)
crew_worker(
  worker = "my_worker",
  store = store$marshal(),
  jobs = 1,
  timeout = 0,
  wait = 0
)
store$read_worker_output("my_worker")$result # 2
}
}
\keyword{internal}
