% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_worker.R
\name{crew_worker}
\alias{crew_worker}
\title{Crew worker.}
\usage{
crew_worker(settings, host, port, token)
}
\arguments{
\item{settings}{Named list of arguments to \code{mirai::server()}.}

\item{host}{IP address of the host to connect to.}

\item{port}{TCP port to register a successful connection to the host.}

\item{token}{Character of length 1 to identify the instance of the
process connected to the socket.}
}
\value{
\code{NULL} (invisibly)
}
\description{
Launch a \code{crew} worker which runs a \code{mirai} server.
}
\seealso{
Other utilities: 
\code{\link{crew_eval}()},
\code{\link{crew_wait}()}
}
\concept{utilities}
