% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_worker.R
\name{crew_worker}
\alias{crew_worker}
\title{Crew worker.}
\usage{
crew_worker(settings, launcher, worker, instance)
}
\arguments{
\item{settings}{Named list of arguments to \code{mirai::server()}.}

\item{launcher}{Character of length 1, name of the launcher.}

\item{worker}{Positive integer of length 1, index of the worker.
This worker index remains the same even when the current instance
of the worker exits and a new instance launches.}

\item{instance}{Character of length 1 to uniquely identify
the current instance of the worker.}
}
\value{
\code{NULL} (invisibly)
}
\description{
Launch a \code{crew} worker which runs a \code{mirai} server.
}
\seealso{
Other utilities: 
\code{\link{crew_assert}()},
\code{\link{crew_eval}()},
\code{\link{crew_random_name}()},
\code{\link{crew_retry}()}
}
\concept{utilities}
