% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_worker_loop.R
\name{crew_worker_loop}
\alias{crew_worker_loop}
\title{Worker event loop.}
\usage{
crew_worker_loop(name, store, timeout, wait_input)
}
\arguments{
\item{name}{Character of length 1, name of the worker.}

\item{store}{Marshaled \code{R6} data store object to
send/receive job data from the crew.
Constructed from the \verb{$marshal()} method of the store.
\code{eval(parse(text = store))}
should reconstruct the usable store object inside the worker event loop.}

\item{timeout}{Positive numeric of length 1, number of seconds
that a worker can idle before timing out.}

\item{wait_input}{Positive numeric of length 1, number of seconds
that the worker waits between checking if a job exists.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Run the worker event loop.
}
\details{
The worker event loop runs inside the underlying process
of a worker. It waits for the next job (timing out
if it idles for too long at one time) runs any incoming jobs
and puts the output in the data store. If sent a shutdown job
(which throws an error of class \code{"crew_shutdown"}) then the
loop exits.
}
\examples{
# Example of send one job that shuts down the loop:
dir_root <- tempfile()
dir.create(dir_root)
store <- class_store_local$new(dir_root = dir_root)
fun <- function(x, store) {
  fun <- function() rlang::abort(message = "x", class = "crew_shutdown")
  data <- list(fun = deparse(fun), args = list())
  store$write_input(name = "my_worker", data = data)
  paste("job", x)
}
args <- list(x = 12, store = store)
data <- list(fun = deparse(fun), args = args)
store$write_input(name = "my_worker", data = data)
crew_worker_loop(
  name = "my_worker",
  store = store$marshal(),
  timeout = Inf,
  wait_input = 0
)
store$read_output("my_worker")$value
}
