% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_worker_loop.R
\name{crew_worker_loop_run}
\alias{crew_worker_loop_run}
\title{Local worker inner event loop.}
\usage{
crew_worker_loop_run(name, store, timeout, wait_input)
}
\arguments{
\item{name}{Character of length 1, name of the worker.}

\item{store}{Marshaled \code{R6} data store object to send/receive job data from the crew.
Constructed from the \verb{$marshal()} method of the store. \code{eval(parse(text = store))}
should reconstruct the usable store object inside the worker event loop.}

\item{timeout}{Positive numeric of length 1, number of seconds
that a worker can idle before timing out.}

\item{wait_input}{Positive numeric of length 1, number of seconds
that the worker waits between checking if a job exists.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Not a user-side function. for internal use only.
}
\keyword{internal}
